### 2021-08-20 5b50204
Simplified models and a controller support
- greatly simplified INS, Stokes, and MCINS
- added support for modelling controllers as part of the simulation.
