+++ models/ins.py
@@ -19,8 +19,7 @@ import ngsolve as ngs
 from helpers.ngsolve_ import get_special_functions
 from helpers.dg import avg, jump, grad_avg
 from models import Model
-from config_functions import ConfigParser
-from typing import List, Optional, Union
+from typing import Dict, List, Optional, Union
 from ngsolve.comp import ProxyFunction
 from ngsolve import Parameter, GridFunction, FESpace, BilinearForm, LinearForm, Preconditioner
 from helpers.error import norm, mean
@@ -31,22 +30,11 @@ class INS(Model):
     A single phase incompressible Navier-Stokes model.
     """
 
-    def __init__(self, config: ConfigParser, t_param: List[Parameter]) -> None:
-        # Specify information about the model components
-        # NOTE: These MUST be set before calling super(), since it's needed in superclass' __init__
-        self.model_components                   = {'u':  0,      'p': 1}
-        self.model_local_error_components       = {'u':  True,   'p': True}
-        self.time_derivative_components         = [{'u': True,   'p': False}]
-        self.num_weak_forms                     = 1
+    def _pre_init(self) -> None:
+        # Nothing needs to be done
+        pass
 
-        # Pre-define which BCs are accepted for this model, all others are thrown out.
-        self.BC_init = {'dirichlet':    {},
-                        'stress':       {},
-                        'parallel':     {},
-                        'pinned':       {}}
-
-        super().__init__(config, t_param)
-
+    def _post_init(self) -> None:
         # Load the solver parameters.
         self.linearize = self.config.get_item(['SOLVER', 'linearization_method'], str)
         if self.linearize not in ['Oseen', 'IMEX']:
@@ -62,6 +50,26 @@ class INS(Model):
                                  'at least 1.')
             self.W = self._construct_linearization_terms()
 
+    def _define_model_components(self) -> Dict[str, Optional[int]]:
+        return {'u': 0,
+                'p': 1}
+
+    def _define_model_local_error_components(self) -> Dict[str, bool]:
+        return {'u': True,
+                'p': True}
+
+    def _define_time_derivative_components(self) -> List[Dict[str, bool]]:
+        return [
+            {'u': True,
+             'p': False}
+        ]
+
+    def _define_num_weak_forms(self) -> int:
+        return 1
+
+    def _define_bc_types(self) -> List[str]:
+        return ['dirichlet', 'stress', 'parallel', 'pinned']
+
     @staticmethod
     def allows_explicit_schemes() -> bool:
         # INS cannot work with explicit schemes
@@ -69,9 +77,21 @@ class INS(Model):
 
     def _set_model_parameters(self) -> None:
         self.kv = self.model_functions.model_parameters_dict['kinematic_viscosity']['all']
-        self.f = self.model_functions.model_functions_dict['source']['all']
+        self.f = self.model_functions.model_functions_dict['source']
 
     def _construct_fes(self) -> FESpace:
+        return ngs.FESpace(self._contruct_fes_helper(), dgjumps=self.DG)
+
+    def _contruct_fes_helper(self) -> List[FESpace]:
+        """
+        Helper function for creating the FESpace.
+
+        This function exists in order to allow multicomponent_ins (and any further additions) to simplify their FES
+        creation by being able to use this helper function to create the finite element space for velocity and pressure.
+
+        Return:
+            A list containing individual finite element spaces
+        """
         if not self.DG:
             if self.element['u'] == 'HDiv' or self.element['u'] == 'RT':
                 print('We recommended that you NOT use HDIV spaces without DG due to numerical issues.')
@@ -84,15 +104,15 @@ class INS(Model):
                              dgjumps=self.DG, RT=True)
         else:
             fes_u = getattr(ngs, self.element['u'])(self.mesh, order=self.interp_ord,
[…]
