+++ solvers/adaptive_transient_solvers/adaptive_two_step.py
@@ -75,8 +75,8 @@ class AdaptiveTwoStep(BaseAdaptiveTransientMultiStepSolver):
         self._assemble()
 
     def _single_solve(self) -> None:
-        self.model.single_iteration(self.a_pred, self.L_pred, self.preconditioner_pred, self.gfu_pred)
-        self.model.single_iteration(self.a_corr, self.L_corr, self.preconditioner_corr, self.gfu)
+        self.model.solve_single_step(self.a_pred, self.L_pred, self.preconditioner_pred, self.gfu_pred)
+        self.model.solve_single_step(self.a_corr, self.L_corr, self.preconditioner_corr, self.gfu)
 
     def _calculate_local_error(self) -> Tuple[List[float], List[float], List[str]]:
         # Include any variables specified by the model as included in local error.
