### 2021-09-21 #3742 @WilkAndy
@nasserma , @elizabethmonte and @Alex-Vasile - i have a question about authorship.  You can read the JOSS guidelines here: https://joss.readthedocs.io/en/latest/submitting.html#authorship .  Given those guidelines, are you OK with James and Nasser being co-authors, even though they appear to have contributed zero code?
