### 2021-09-21 #3742 @WilkAndy
@nasserma , since i am a heavy contributor to [MOOSE](https://mooseframework.org) i don't like your statement "Open-source packages ... such as MOOSE ... have very steep learning curves and are inaccessible to the general
scientific and engineering communities."   MOOSE was designed to be easy to use.  Perhaps it has changed into something more complicated over time.  Anyway, your statement is just your opinion.  Feel free to make no changes, but I guess i'd tone it down a little, by perhaps changing to "... can have steep learning curves." (without the "inaccessible" part).
