### 2021-10-05 #3742 @nasserma
@lucydot  @WilkAndy This is our first time submitting to JOSS, I just want to confirm how we respond to reviewer comments. Should we be collating them and responding after they are complete or should we respond to each one individually? We have a response ready for @WilkAndy but it also involves revisions to the manuscript.
