### 2021-10-07 #3742 @WilkAndy
Just trying to install and use.  Please add instructions for getting "NGSolve version 6.2.2101" to your "IO" docs
