### 2021-10-07 #3742 @nasserma
> @nasserma , since i am a heavy contributor to [MOOSE](https://mooseframework.org) i don't like your statement "Open-source packages ... such as MOOSE ... have very steep learning curves and are inaccessible to the general scientific and engineering communities." MOOSE was designed to be easy to use. Perhaps it has changed into something more complicated over time. Anyway, your statement is just your opinion. Feel free to make no changes, but I guess i'd tone it down a little, by perhaps changing to "... can have steep learning curves." (without the "inaccessible" part).

We have relaxed the language in the revised manuscript, as suggested. Our intention was never to diminish other software packages' utility, but rather differentiate OpenCMP's purpose. Our core mandate is to enable users who are familiar with the command-line interface to perform simulations without having to do any programming whatsoever, while simultaneously empowering developers to contribute new multiphysics models.
