### 2021-10-07 #3742 @nasserma
> @nasserma , i have not yet read your paper in detail, nor downloaded the software, played with it, etc, but overall i think this paper should be significantly longer. You may have satisfied the minimum requirements for a JOSS paper, but i feel to excite readers to try your software, you need more details. I think you should mention:
>
>     * code coverage
>
>     * perhaps outline the tutorials
>
>     * a bit more details about the user interface
>
>     * a picture of an example, perhaps explaining why your code/approach is advantageous in terms of usability, accuracy, computational efficiency, etc, compared with other codes.  Sometimes it is hard to not denigrate excellent alternative codes when you're doing this, but you needn't go over-the-top: you're just trying to show the reader your code is nice.
>
>
> This is really summarizing work you've already done, so i hope it won't be a huge chore for you.

Based on the JOSS requirements for submissions, "the paper should be between 250-1000 words" and the submitted manuscript is at the upper limit. We could add more details into the paper, but it would be at the expense of other text, most of which has been added to meet the requirements of what the paper should include from the JOSS website,

https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements

We have added references to the OpenCMP website in the revised manuscript. The website content addresses all of your comments except for code coverage, which we have added to the revised manuscript without exceeding the 1000 word limit.
