### 2021-10-07 #3742 @Alex-Vasile
> @nasserma , @elizabethmonte and @Alex-Vasile - i have a question about authorship. You can read the JOSS guidelines here: https://joss.readthedocs.io/en/latest/submitting.html#authorship . Given those guidelines, are you OK with James and Nasser being co-authors, even though they appear to have contributed zero code?

First, a note about code contributions to the project. The project began on a private internal repo, with commits being periodically being rolled up and copied into the GitHub repo. This results in actual code contributions being unintentionally obscured on the GitHub repo. Future commits will better reflect authorship as development branches are switched from internal repos to GitHub.

To address your question directly, yes I am okay with James and Nasser being co-authors given that they have been active participants. James has contributed code to the repository in the past and is actively implementing new models at the moment. Nasser, as well as James, has provided direction in terms of code structure, and help in understanding, implementing, debugging the numerical methods.
