### 2021-10-11 #3742 @bonh
Hi @nasserma!

I am assessing the scholar effort of your code. In my opinion the performance of simulation codes is a very important feature (the most sophisticated implementation of a wonderful physical model will not have a wider impact if the solution time is unacceptable). About the performance of your code (haven't tested it yet): Is it possible to run your code in parallel? In the paper you mention, that utilizing MPI is planned but are you able to use shared memory? Have you compared the performance of your code against other codes? Do you have performed three-dimensional simulations?
