### 2021-11-15 uw-comphys/opencmp#5 @Alex-Vasile
JOSS: Add Installation Instructions for NGSolve
As per https://github.com/openjournals/joss-reviews/issues/3742#issuecomment-937444021.
