### 2021-11-15 uw-comphys/opencmp#6 @Alex-Vasile
JOSS: Performance of code
1. Is it possible to run your code in parallel?
2. In the paper you mention, that utilizing MPI is planned but are you able to use shared memory?
3. Have you compared the performance of your code against other codes?
4. Do you have performed three-dimensional simulations?"

per: https://github.com/openjournals/joss-reviews/issues/3742#issuecomment-940420033
