### 2021-11-15 uw-comphys/opencmp#7 @Alex-Vasile
JOSS: Diffuse Interface Method
1. arxiv "article shares a lot of the authors but the software OpenCMP is not mentioned. Did you use a different implementation there? If not, I think it would be a good idea to cross-reference your papers. In the present joss paper a clear reference to a more physics and simulation results focused paper would help the interested reader to find details of your work."
2. "using the diffuse interface method to represent immersed boundaries is not a new idea presented for the first time in arxiv:2101.06824, right? Perhaps you should add a reference to the paper which indicates this and use your reference to refer the reader to the improvements you made to the method."
3. "You mention that there are no other implementations of the diffuse interface method. Is that really correct? How about the Phase Field implementation of comsol? Can that not be used to simulate "hard" boundaries by choosing super high values for density and viscosity or so?"

per: https://github.com/openjournals/joss-reviews/issues/3742#issuecomment-940427034
