### 2021-11-15 uw-comphys/opencmp#8 @Alex-Vasile
JOSS: Expand on paper
1. code coverage
2. perhaps outline the tutorials
3. a bit more details about the user interface
4. a picture of an example, perhaps explaining why your code/approach is advantageous in terms of usability, accuracy, computational efficiency, etc, compared with other codes. Sometimes it is hard to not denigrate excellent alternative codes when you're doing this, but you needn't go over-the-top: you're just trying to show the reader your code is nice.

per: https://github.com/openjournals/joss-reviews/issues/3742#issuecomment-924368594

Go over 1000 words to address them in more detail within the paper itself, per https://github.com/openjournals/joss-reviews/issues/3742#issuecomment-940430141.
