### 2021-12-01 uw-comphys/opencmp#5 @WilkAndy
Prob can just link to https://docu.ngsolve.org/latest/install/install_sources.html and/or https://ngsolve.org/downloads
