### 2021-12-01 uw-comphys/opencmp#7 @WilkAndy
Monte(2021) is published: add the DOI for https://onlinelibrary.wiley.com/doi/epdf/10.1002/cjce.24320 to the JOSS paper
