### 2021-12-01 uw-comphys/opencmp#9 @WilkAndy
Define "Multi-component reacting NS" in JOSS paper
This could mean different things to different people.  Eg, geochemists may expect full geochemical ability, while petroleum researchers may expect the ability to study reservoir stimulation, while nuclear people would want radionuclide decay.

As per https://github.com/openjournals/joss-reviews/issues/3742
