### 2021-12-01 #3742 @WilkAndy
Comment: define your physics in Table(1) through equations .  Eg "multi-component reacting NS" might mean lots of things.

Refs https://github.com/uw-comphys/opencmp/issues/9
