### 2021-12-01 uw-comphys/opencmp#9 @WilkAndy
Also should probably write what you mean by "stokes", "poisson", etc, for completeness.
