### 2021-12-02 #3742 @WilkAndy
Just trying to install on a fresh windows machine.  Also need:
```
pip install scipy
```
