### 2021-12-02 #3742 @WilkAndy
I think automated tests are mandatory, and i can't find any.  You have tutorial examples - great, a new user can test things are working to some degree, but it wouldn't take much work to build a test suite out of those, so that new users can be confident things are working as planned.  For instance in "tutorial 4", various norms are outputted, but these are different for my computer than are shown at https://opencmp.io/tutorials/tutorial_4.html .   While i'm not too concerned at the differences, it would be much easier and confidence-inspiring if new users could just write
```
python run_all_tests.py
```
and OpenCMP replies back: "all tests passed".  I use the python `unittest`.

Refs https://github.com/uw-comphys/opencmp/issues/11
