### 2021-12-02 uw-comphys/opencmp#12 @WilkAndy
tutorial 5 fails
I'm running this on a clean windows machine, so perhaps i need to `pip install` something.  My error is:
```
importing NGSolve-6.2.2104
Using the default value of False for CONTROLLER, active.
Traceback (most recent call last):
  File "run.py", line 112, in <module>
    main(sys.argv[1])
  File "run.py", line 62, in main
    solver = solver_class(model_class, config)
  File "C:\Users\asdf\Documents\opencmp-main\solvers\transient_multistep.py", line 38, in __init__
    super().__init__(model_class, config)
  File "C:\Users\asdf\Documents\opencmp-main\solvers\base_solver.py", line 181, in __init__
    raise ValueError('IMEX linearization must be used with an IMEX time integration scheme.')
ValueError: IMEX linearization must be used with an IMEX time integration scheme.
```
