### 2021-12-02 uw-comphys/opencmp#14 @WilkAndy
tutorial 7 fails
Error is:
```
importing NGSolve-6.2.2104
Using the default value of False for ERROR ANALYSIS, check_error.
Using the default value of False for CONTROLLER, active.
Traceback (most recent call last):
  File "run.py", line 112, in <module>
    main(sys.argv[1])
  File "run.py", line 62, in main
    solver = solver_class(model_class, config)
  File "C:\Users\asdf\Documents\opencmp-main\solvers\transient_multistep.py", line 38, in __init__
    super().__init__(model_class, config)
  File "C:\Users\asdf\Documents\opencmp-main\solvers\base_solver.py", line 151, in __init__
    self.model = model_class(self.config, self.t_param)
  File "C:\Users\asdf\Documents\opencmp-main\models\base_model.py", line 184, in __init__
    self.ic_functions.set_initial_conditions(self.IC, self.mesh, self.name, self.model_components)
  File "C:\Users\asdf\Documents\opencmp-main\config_functions\initial_conditions.py", line 78, in set_initial_conditions
    val = self._find_rel_path_for_file(val)
  File "C:\Users\asdf\Documents\opencmp-main\config_functions\base_config_functions.py", line 83, in _find_rel_path_for_file
    raise FileNotFoundError('The given file does not exist.')
FileNotFoundError: The given file does not exist.
```
