### 2021-12-02 #3742 @WilkAndy
Various tutorials fail to run.  These could be because i'm running on a vanilla windows system with no extra packages installed.

Refs https://github.com/uw-comphys/opencmp/issues/12
Refs https://github.com/uw-comphys/opencmp/issues/13
Refs https://github.com/uw-comphys/opencmp/issues/14
Refs https://github.com/uw-comphys/opencmp/issues/15
