### 2021-12-02 #3742 @WilkAndy
I haven't read the documentation carefully, I'm sorry.  However, at the moment i believe all your tutorials are "simple" cases that can be solved easily using other software.  If true, this doesn't really compel me to try your new software.  I believe the big new thing about your software is diffuse boundary method.  How about highlighting that through an awesome example, both in your online documentation and this JOSS paper.
