### 2021-12-02 #3742 @nasserma
> I think automated tests are mandatory, and i can't find any. You have tutorial examples - great, a new user can test things are working to some degree, but it wouldn't take much work to build a test suite out of those, so that new users can be confident things are working as planned. For instance in "tutorial 4", various norms are outputted, but these are different for my computer than are shown at https://opencmp.io/tutorials/tutorial_4.html . While i'm not too concerned at the differences, it would be much easier and confidence-inspiring if new users could just write
>
> ```
> python run_all_tests.py
> ```
>
> and OpenCMP replies back: "all tests passed". I use the python `unittest`.
>
> Refs [uw-comphys/opencmp#11](https://github.com/uw-comphys/opencmp/issues/11)

Please see the README.md file which describes our unit test procedure. We use pytest instead of unittest for unit and integration tests. We prefer it for a few reasons, but if you see the pytests/ directory there is also a README.md file with further information regarding how (i) how run the existing tests and (ii) how to create new tests.

At this point we have quite a decent amount of tests, they are all present in subdirectories of pytests/ and are commented/documented.

That being said, the most recent push to the Github branch seems to have caused some significant issues with pytests and the tutorials. We are in the process of working out how to deal with our internal VCS and external VCS (Github). I will work this out with the others ASAP.
