### 2021-12-03 #3742 @nasserma
> Various tutorials fail to run. These could be because i'm running on a vanilla windows system with no extra packages installed.
>
> Refs [uw-comphys/opencmp#12](https://github.com/uw-comphys/opencmp/issues/12) Refs [uw-comphys/opencmp#13](https://github.com/uw-comphys/opencmp/issues/13) Refs [uw-comphys/opencmp#14](https://github.com/uw-comphys/opencmp/issues/14) Refs [uw-comphys/opencmp#15](https://github.com/uw-comphys/opencmp/issues/15)

Please use revision ID  a475159 for now, there was definitely an issue pushing changes from the last revision and we will resolve in shortly.
