### 2021-12-05 9e7954c
Update stokes.py
Reverting changes to Stokes. Closes #12 #13
