### 2021-12-05 8308a97
Update requirements
Add requirements.txt file, update installation instructions to have the most up-to-date list of required packages.

Closes #10
