+++ requirements.txt
@@ -1,0 +1,7 @@
+edt
+meshio
+numpy
+pyparsing
+pytest>=6.2
+scipy
+tabulate
