### 2021-12-05 uw-comphys/opencmp#14 @Alex-Vasile
It appears that it's having trouble finding the initial condition. (Failing in `initial_conditions.py`).

Tutorials 7 and 8 feature two config files, `config` and `config_IC`.

`config_IC` creates the initial condition files used by `config` for the main solve.

Is this crash happening when you run opencmp with `config` or with `config_IC`?
