### 2021-12-05 uw-comphys/opencmp#11 @Alex-Vasile
A range on python unit tests are available under `pytests/`. They can be run with `pytest pytests/`.

**NOTE:** At the moment, running them will require adding the main OpenCMP folder to your python path, e.g. running `echo PYPOTHONPATH="path_to_opencmp_main_folder:$PYTHONPATH"` in the terminal first.

The tests are a mixture of testing individual functions and modules, as well as full simulations. As such, they take a non-trivial amount of time to run, about 15 minutes to run on a 4 core, 8 thread, desktop machine.
