### 2021-12-09 uw-comphys/opencmp#6 @nasserma
@bonh It is currently possible to run OpenCMP using shared-memory parallelization (numthreads parameter in config file). The underlying finite element library, NGSolve, also supports distributed memory parallelization using MPI, but OpenCMP currently does not support this usage. We are planning this in the future (2022).

We have not tested the performance of OpenCMP against other packages, however, the underlying finite element library has a performance analysis found [here]([url](https://www.asc.tuwien.ac.at/~schoeberl/wiki/publications/ngs-cpp11.pdf)).

We have performed 3D simulations using shared memory parallelization, we are creating a tutorial demonstrating this shortly.
