### 2021-12-09 uw-comphys/opencmp#9 @nasserma
We have modified the paper to indicate that the Poisson equation can also be referred to as the Heat equation. Regarding Stokes equations, this is a standard reference to the incompressible Navier-Stokes (INS) equations in the limit of low Reynolds number, so we do not think further clarification is needed and it is not provided in the documentation of other software packages that include it in their examples.

We have changed our naming of the last model to "Multicomponent Mixture INS equations" for clarity. We support inclusion support for first order reactions and are in the progress of documenting support for higher order reactions (nonlinear). Regarding petroleum researchers, they would expect reference to transport in porous media, and we are not very familiar with what you mean about radionuclide decay.
