### 2021-12-10 869198b
Adding 3D DIM example.
