### 2021-12-12 99a2d1a
Turned the DIM example into tutorial 10 and added visualization with matplotlib.
