### 2021-12-13 uw-comphys/opencmp#6 @bonh
In my opinion some sort of shared memory parallelization is mandatory for modern CFD tools (can you think of any tool without MPI or something similar so that it can run on a supercomputer?). Otherwise the tool might just be a great educational package (which is not bad per se, but I do not think this is your goal, right?). So perhaps it might be a good idea to include the MPI capabilities of NGSolve into OpenCMP before publishing a paper in JOSS, don't you think? However, I am really looking forward to the 3D simulation example using shared memory parallelization!
