### 2021-12-13 #3742 @bonh
I will continue my review after we have discussed uw-comphys/opencmp#6 and uw-comphys/opencmp#7. I rate these as critical to assess the scholarly effort.
