### 2021-12-13 uw-comphys/opencmp#6 @bonh
Oh, sry, I meant "distributed memory":
> In my opinion some sort of **distributed memory** parallelization is mandatory for modern CFD tools
