### 2021-12-13 uw-comphys/opencmp#6 @nasserma
> Oh, sry, I meant "distributed memory":
>
> > In my opinion some sort of **distributed memory** parallelization is mandatory for modern CFD tools

OK, then from my understanding you are requiring use to implement distributed-memory parallelization for this submission?
