### 2021-12-13 uw-comphys/opencmp#6 @bonh
By no means "requiring"! But you wrote yourself that you plan to implement it in 2022. And I just think that is a great idea! So I am merely suggesting that the implementation of a distributed parallelization would greatly enhance the publication. Otherwise I would need to look for the 3D simulations. But just to be clear, in my opinion, a CFD package which is not able to at least simulate moderate scale problems (which you have to demonstrate as of yer, distributed or shared memory) AND does not implement some new methodology (see the discussion of the diffuse interface method) does not meet the "scholarly effort" described by JOSS.
