### 2021-12-13 #3742 @bonh
I am sorry for not being more clear: recommending the paper for publication in JOSS and improving the package are two different parts of my work here, I think. So, as I wrote just now in uw-comphys/opencmp#6,

> in my opinion, a CFD package which is not able to at least simulate moderate scale problems AND does not implement some new methodology does not meet the "scholarly effort" described by JOSS.

I just want to engage in a discussion with you about your work! Now if you are able to simulate moderate size problems (e.g., something in 3D or so, using shared memory) this would be okay for me to proceeded with the review. Same way if you convince me of the novelty of your method (diffuse interface).
However, I think that if you offer the user MPI (or some other distributed memory parallelization), this would greatly improve your software and therefore the paper!
