### 2021-12-14 e722420
- Changed source file structure for use with setuputils as a build environment. - Implemented changes for using setuputils so that pip can be used to install. - Added diffuse interface with tutorial 10 and documentation. - Added preliminary integrated controller functionality, with pytests, and an example. - Various minor fixes.
