+++ publications/2108_JOSS/paper.md
@@ -9,7 +9,7 @@ tags:
  - diffuse interface method
  - immersed boundary method
 authors:
- - name: Elizabeth J. Monte
+ - name: Elizabeth Julia Monte
    orcid: 0000-0001-7328-775X
    affiliation: 1
  - name: Alexandru Andrei Vasile
@@ -34,7 +34,7 @@ bibliography: paper.bib
 
 OpenCMP is a computational multiphysics software package based on the finite element method [@Ferziger2002]. It is primarily intended for physicochemical processes in which fluid convection plays a significant role. OpenCMP uses the NGSolve finite element library [@ngsolve] for spatial discretization and provides a configuration file-based interface for pre-implemented models and time discretization schemes. It also integrates with Netgen [@ngsolve] and Gmsh [@Geuzaine2009] for geometry construction and meshing. Additionally, it provides users with built-in functionality for post-processing, error analysis, and data export for visualisation using Netgen [@ngsolve] or ParaView [@Ahrens2005]. 
 
-OpenCMP development follows the principles of ease of use, performance, and extensibility. The configuration file-based user interface is intended to be concise, readable, and intuitive. Furthermore, the code base is structured such that experienced users with appropriate background can add their own models with minimal modifications to existing code. The finite element method enables the use of high-order polynomial interpolants for increased simulation accuracy, however, continuous finite element methods suffer from stability and accuracy (conservation) for fluid convection-dominated problems. OpenCMP addresses this by providing discontinuous Galerkin method [@Cockburn2000] solvers, which are locally conservative and improve simulation stability for convection-dominated problems. Finally, OpenCMP implements the diffuse interface method [@Monte2021], an immersed boundary method [@Mittal2005], which allows complex simulation domains to be meshed by non-conforming structured meshes for improved simulation stability and reduced computational complexity (under certain conditions [@Monte2021]).
+OpenCMP development follows the principles of ease of use, performance, and extensibility. The configuration file-based user interface is intended to be concise, readable, and intuitive. Furthermore, the code base is structured such that experienced users with appropriate background can add their own models with minimal modifications to existing code. The finite element method enables the use of high-order polynomial interpolants for increased simulation accuracy, however, continuous finite element methods suffer from stability and accuracy (conservation) for fluid convection-dominated problems. OpenCMP addresses this by providing discontinuous Galerkin method [@Cockburn2000] solvers, which are locally conservative and improve simulation stability for convection-dominated problems. Finally, OpenCMP implements the diffuse interface method [@Monte2021,@Nguyen2018], an immersed boundary method [@Mittal2005], which allows complex simulation domains to be meshed by non-conforming structured meshes for improved simulation stability and reduced computational complexity (under certain conditions [@Monte2021]).
 
 # Statement of Need
 
@@ -52,14 +52,14 @@ The table below summarises the current capabilities of OpenCMP. Future work on O
 
 | Feature           | Description                                                  |
 | ----------------- | ------------------------------------------------------------ |
-| Meshing           | Accepts Netgen [@ngsolve] or Gmsh [@Geuzaine2009] meshes      |
+| Meshing           | Accepts Netgen [@ngsolve] or Gmsh [@Geuzaine2009] meshes     |
 | Numerical Methods | Standard continuous Galerkin finite element method           |
 |                   | Discontinuous Galerkin finite element method                 |
 |                   | Diffuse interface method                                     |
-| Models            | Poisson equation                                             |
+| Models            | Poisson (Heat) equation                                      |
 |                   | Stokes equations                                             |
-|                   | Single-component incompressible Navier-Stokes equations      |
-|                   | Multi-component reacting incompressible Navier-Stokes equations |
+|                   | Incompressible Navier-Stokes (INS) equations                 |
+|                   | Multicomponent Mixture INS equations                         |
[…]
