+++ publications/2108_JOSS/paper.bib
@@ -48,10 +48,12 @@
 }
 
 @Unpublished{Monte2021,
-  author = {E. J. Monte and J. Lowman and N. M. Abukhdeir},
-  title  = {A diffuse interface method for simulation-based screening of heat transfer processes with complex geometries},
-  note   = {arXiv:2101.06824},
-  year   = {2021},
+  title={A diffuse interface method for simulation-based screening of heat transfer processes with complex geometries},
+  author={Monte, Elizabeth J and Lowman, James and Abukhdeir, Nasser Mohieddin},
+  journal={The Canadian Journal of Chemical Engineering},
+  year={2021},
+  publisher={Wiley Online Library},
+  doi	   = {10.1002/cjce.24320},
 }
 
 @Article{Mittal2005,
