+++ opencmp/models/poisson.py
@@ -16,11 +16,10 @@
 ########################################################################################################################
 
 import ngsolve as ngs
-from helpers.ngsolve_ import get_special_functions
-from helpers.dg import jump, grad_avg
-from config_functions import ConfigParser
-from models import Model
-from typing import Tuple, List, Union, Optional
+from ..helpers.ngsolve_ import get_special_functions
+from ..helpers.dg import jump, grad_avg
+from . import Model
+from typing import Dict, List, Union, Optional
 from ngsolve.comp import FESpace, ProxyFunction
 from ngsolve import Parameter, GridFunction, BilinearForm, LinearForm, Preconditioner
 
@@ -30,28 +29,39 @@ class Poisson(Model):
     A poisson model.
     """
 
-    def __init__(self, config: ConfigParser, t_param: List[Parameter]) -> None:
-        # Specify information about the model components
-        # NOTE: These MUST be set before calling super(), since it's needed in superclass' __init__
-        self.model_components               = {'u':  None}
-        self.model_local_error_components   = {'u':  True}
-        self.time_derivative_components     = [{'u': True}]
-        self.num_weak_forms                 = 1
+    def _pre_init(self) -> None:
+        # Nothing needs to be done
+        pass
 
-        # Pre-define which BCs are accepted for this model, all others are thrown out.
-        self.BC_init = {'robin':     {},
-                        'dirichlet': {},
-                        'neumann':   {}}
+    def _post_init(self) -> None:
+        # Nothing needs to be done
+        pass
 
-        super().__init__(config, t_param)
+    def _define_model_components(self) -> Dict[str, Optional[int]]:
+        return {'u': 0}
+
+    def _define_model_local_error_components(self) -> Dict[str, bool]:
+        return {'u': True}
+
+    def _define_time_derivative_components(self) -> List[Dict[str, bool]]:
+        return [{'u': True}]
+
+    def _define_num_weak_forms(self) -> int:
+        return 1
+
+    def _define_bc_types(self) -> List[str]:
+        return ['robin', 'dirichlet', 'neumann']
 
     @staticmethod
     def allows_explicit_schemes() -> bool:
         return True
 
     def _construct_fes(self) -> FESpace:
-        return getattr(ngs, self.element['u'])(self.mesh, order=self.interp_ord,
+        fes = getattr(ngs, self.element['u'])(self.mesh, order=self.interp_ord,
                                              dirichlet=self.dirichlet_names.get('u', ''), dgjumps=self.DG)
+        compound_fes = ngs.FESpace([fes])
+
+        return compound_fes
 
     def _set_model_parameters(self) -> None:
         self.dc = self.model_functions.model_parameters_dict['diffusion_coefficient']['all']
@@ -72,8 +82,8 @@ class Poisson(Model):
 
         return [a]
 
-    def construct_bilinear_time_coefficient(self, U: List[ProxyFunction], V: List[ProxyFunction],
-                                    dt: Parameter = Parameter(1.0), time_step: int = 0) -> List[BilinearForm]:
+    def construct_bilinear_time_coefficient(self, U: List[ProxyFunction], V: List[ProxyFunction], dt: Parameter,
+                                            time_step: int) -> List[BilinearForm]:
         # Split up the trial (weighting) and test functions
         u = U[0]
         v = V[0]
@@ -87,8 +97,8 @@ class Poisson(Model):
 
         return [a]
 
-    def construct_linear(self, V: List[ProxyFunction], gfu_0: Optional[List[GridFunction]] = None,
-                         dt: Parameter = Parameter(1.0), time_step: int = 0) -> List[LinearForm]:
[…]
