+++ opencmp/solvers/adaptive_transient_solvers/base_adaptive_transient_RK.py
@@ -17,12 +17,12 @@
 
 import math
 # TODO: generalize so we don't have to import each one individually
-from models import Model
+from ...models import Model
 from typing import Tuple, Type, List
-from config_functions import ConfigParser
-from solvers import TransientRKSolver
+from ...config_functions import ConfigParser
+from ...solvers import TransientRKSolver
 from abc import ABC, abstractmethod
-from helpers.saving import SolutionFileSaver
+from ...helpers.saving import SolutionFileSaver
 import sys
 
 """

[…]
