### 2021-12-14 98be2eb
- Deleted several files that were not removed from previous merge. - Moved requirements into setup.cfg.
