+++ controllers/pid_controller.py
@@ -1,174 +1,0 @@
-########################################################################################################################
-# Copyright 2021 the authors (see AUTHORS file for full list).                                                         #
-#                                                                                                                      #
-# This file is part of OpenCMP.                                                                                        #
-#                                                                                                                      #
-# OpenCMP is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public  #
-# License as published by the Free Software Foundation, either version 2.1 of the License, or (at your option) any     #
-# later version.                                                                                                       #
-#                                                                                                                      #
-# OpenCMP is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied        #
-# warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more  #
-# details.                                                                                                             #
-#                                                                                                                      #
-# You should have received a copy of the GNU Lesser General Public License along with OpenCMP. If not, see             #
-# <https://www.gnu.org/licenses/>.                                                                                     #
-########################################################################################################################
-
-from .base_controller import Controller
-from models import Model
-from ngsolve import CoefficientFunction, GridFunction, Parameter
-from typing import Dict, List, Optional, Union, cast
-from numpy import isclose
-from helpers.math import H_s, H_t
-
-
-class PID(Controller):
-    """
-    A Discrete PID controller.
-    """
-
-    def __init__(self, t_params: List[Parameter], model: Model, config_rel_path: str, import_dir: str) -> None:
-        super().__init__(t_params, model, config_rel_path, import_dir)
-
-        # Ensure that this is a Single-Input-Single-Output controller
-        assert len(self.vars_control)    == len(t_params)
-        assert len(self.vars_manipulate) == len(t_params)
-
-        # Get information for manipulated variable
-        self.bc_type     = self.vars_manipulate[0][0]
-        self.bc_var      = self.vars_manipulate[0][1]
-        self.bc_location = self.vars_manipulate[0][2]
-
-        # Extract parameters
-        parameters:  Dict[str, List[float]]   = cast(Dict[str, List[float]],
-                                                     self.config_func.config.get_one_level_dict('PARAMETERS',
-                                                                                                import_dir,
-                                                                                                model.mesh,
-                                                                                                [self.t_params[0]])[0])
-        self.bias:   float              = parameters['bias'][0]
-        self.K_c:    float              = parameters['k_c'][0]
-        self.tau_I:  float              = parameters['tau_i'][0]
-        self.tau_D:  float              = parameters['tau_d'][0]
-
-        # List of past actions, needed for Integral and Derivative control
-        # Initialize with initial error
-        self.errors: List[float] = []
-
-        # The integrated error used for the Integral portion of the controller
-        self.integrated_error = 0.0
-
-        # Load the initial BC value as the initial value to ensure that there's always a smooth transition
-        # TODO: How do we handle it when the initial value is a coefficient function?
-        initial_bc_val = self.model.BC.get(self.bc_type, {}).get(self.bc_var, {}).get(self.bc_location, None)[0]
-        # TODO: when would this be none?
-        if initial_bc_val is not None:
-            if isinstance(initial_bc_val, float):
-                self.prev_control_action_val = cast(float, initial_bc_val)
-            else:
[…]
