+++ controllers/__init__.py
@@ -1,28 +1,0 @@
-########################################################################################################################
-# Copyright 2021 the authors (see AUTHORS file for full list).                                                         #
-#                                                                                                                      #
-# This file is part of OpenCMP.                                                                                        #
-#                                                                                                                      #
-# OpenCMP is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public  #
-# License as published by the Free Software Foundation, either version 2.1 of the License, or (at your option) any     #
-# later version.                                                                                                       #
-#                                                                                                                      #
-# OpenCMP is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied        #
-# warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more  #
-# details.                                                                                                             #
-#                                                                                                                      #
-# You should have received a copy of the GNU Lesser General Public License along with OpenCMP. If not, see             #
-# <https://www.gnu.org/licenses/>.                                                                                     #
-########################################################################################################################
-
-# Superclass
-from .base_controller import Controller
-
-# Implemented controllers
-from .pid_controller import PID
-
-# Controller group
-from .controller_group import ControllerGroup
-
-# Helper function
-from .misc import get_controller
