+++ setup.cfg
@@ -18,6 +18,12 @@ scripts =
 install_requires =
     numpy
     scipy
+    edt
+    meshio
+    pyparsing
+    pytest>=6.2
+    tabulate
+
 
 [options.package_data]
 

[…]
