### 2021-12-14 uw-comphys/opencmp#14 @nasserma
Following-up on Alex's comment, please see the tutorial instructions,

https://opencmp.io/tutorials/tutorial_7.html

you need to solve for the initial condition first.
