### 2021-12-14 uw-comphys/opencmp#15 @nasserma
Please see the tutorial instructions,

https://opencmp.io/tutorials/tutorial_8.html

you need to solve for the initial condition first.
