### 2021-12-14 uw-comphys/opencmp#11 @nasserma
Following-up on Alex's comment, unit tests using the pytest module were provided and may be run using that module from the top-level of the installation. I have added more detail to the installation guide to clarify this,

https://opencmp.io/getting_started/installation_guide.html
