### 2021-12-14 uw-comphys/opencmp#6 @nasserma
We have added a tutorial to show how OpenCMP can be used to solve a 3D problem (heat conduction) with the diffuse interface method, instead of a conformal mesh, for a relatively complex geometry. The tutorial is provided here,

https://opencmp.io/tutorials/tutorial_10.html

We have not added the example to the paper, but if this is desirable please let us know and we will add a section "showing off" the diffuse interface method.
