### 2021-12-14 uw-comphys/opencmp#5 @nasserma
Please see the updated installation instructions, this change has been made,

https://opencmp.io/getting_started/installation_guide.html
