### 2021-12-14 uw-comphys/opencmp#8 @nasserma
We have address comments 1-3 in our most recent revision. Regarding comment #4, we are confirming with one of the reviewers regarding what example would be best for this additional section.
