### 2021-12-14 uw-comphys/opencmp#7 @nasserma
Responding to these points in order:

1. We used a different implementation in Monte(2021), which preceded the one in OpenCMP, which we have clarified in the most recent revision of the paper, along with updating the reference indicated publication.
2.  We added an additional reference for the diffuse interface method (with stabilization) that we used Nguyen(2018). Note that a review article on immersed boundary methods is also included Mittal(2005).
3. To our knowledge, this is the only open-source implementation of the diffuse interface method with stabilization for Dirichlet boundary conditions (see Nguyen(2018)). What you are describing with Comsol's phase field model would not straightforward to implement, would not be stable when using Dirichlet BCs, and is not open-source.
