### 2021-12-14 #3742 @nasserma
@bonh @WilkAndy @lucydot Please see the recent revision that we pushed, it addresses many of the comments that were made, although there are a few outstanding issues that I would like to discuss before we address them.

Please see the open issues,

https://github.com/uw-comphys/opencmp/issues

and our responses. Once we have your feedback we'll make the next round of revisions and please note that the installation instructions have changed,

https://opencmp.io/getting_started/installation_guide.html
