### 2021-12-15 3fa1bed
- Corrected setup.cfg to recurse into the opencmp/ directory and install all subpackages. - Modified the installation instructions to install normally and not in editable mode. - Added a missing bibliography entry for the JOSS publication.
