+++ publications/2108_JOSS/paper.bib
@@ -1,3 +1,14 @@
+@Article{Nguyen2018,
+  author    = {Nguyen, Lam H. and Stoter, Stein K.F. and Ruess, Martin and Sanchez Uribe, Manuel A. and Schillinger, Dominik},
+  journal   = {Int. J. Numer. Methods Eng.},
+  title     = {The diffuse {N}itsche method: Dirichlet constraints on phase-field boundaries},
+  year      = {2018},
+  number    = {4},
+  pages     = {601-633},
+  volume    = {113},
+  doi       = {10.1002/nme.5628},
+}
+
 @Book{Ferziger2002,
   title     = {Computational Methods for Fluid Dynamics},
   publisher = {Springer-Verlag},
