+++ pytests/full_system/ins/test_ins.py
@@ -138,7 +138,8 @@ class TestTransient:
     #   # Run
     #   automated_output_check(capsys, sinusoidal_transient, [2.5e-3, 1e-3])
 
-    def test_sinusoidal_imex_euler_cg(self, capsys: CaptureFixture, sinusoidal_transient: ConfigParser) -> None:
+    def test_sinusoidal_imex_euler_cg(self, capsys: CaptureFixture,
+                                                sinusoidal_transient: ConfigParser) -> None:
         # Change linearization scheme
         sinusoidal_transient['SOLVER']['linearization_method'] = 'IMEX'
         # Change time discretization scheme
@@ -146,7 +147,8 @@ class TestTransient:
         # Run
         automated_output_check(capsys, sinusoidal_transient, [1e-4, 2e-3])
 
-    def test_sinusoidal_imex_euler_dg(self, capsys: CaptureFixture, sinusoidal_transient: ConfigParser) -> None:
+    def test_sinusoidal_imex_euler_dg(self, capsys: CaptureFixture,
+                                                sinusoidal_transient: ConfigParser) -> None:
         # Change linearization scheme
         sinusoidal_transient['SOLVER']['linearization_method'] = 'IMEX'
         # Change time discretization scheme
@@ -158,7 +160,8 @@ class TestTransient:
         # Run
         automated_output_check(capsys, sinusoidal_transient, [1e-4, 5e-11])
 
-    def test_sinusoidal_imex_cnlf_cg(self, capsys: CaptureFixture, sinusoidal_transient: ConfigParser) -> None:
+    def test_sinusoidal_imex_cnlf_cg(self, capsys: CaptureFixture,
+                                                sinusoidal_transient: ConfigParser) -> None:
         # Change linearization scheme
         sinusoidal_transient['SOLVER']['linearization_method'] = 'IMEX'
         # Change time discretization scheme
@@ -166,7 +169,8 @@ class TestTransient:
         # Run
         automated_output_check(capsys, sinusoidal_transient, [1e-4, 2e-3])
 
-    def test_sinusoidal_imex_cnlf_dg(self, capsys: CaptureFixture, sinusoidal_transient: ConfigParser) -> None:
+    def test_sinusoidal_imex_cnlf_dg(self, capsys: CaptureFixture,
+                                                sinusoidal_transient: ConfigParser) -> None:
         # Change linearization scheme
         sinusoidal_transient['SOLVER']['linearization_method'] = 'IMEX'
         # Change time discretization scheme
@@ -178,7 +182,8 @@ class TestTransient:
         # Run
         automated_output_check(capsys, sinusoidal_transient, [2.5e-3, 1e-3])
 
-    def test_sinusoidal_imex_sbdf_cg(self, capsys: CaptureFixture, sinusoidal_transient: ConfigParser) -> None:
+    def test_sinusoidal_imex_sbdf_cg(self, capsys: CaptureFixture,
+                                                sinusoidal_transient: ConfigParser) -> None:
         # Change linearization scheme
         sinusoidal_transient['SOLVER']['linearization_method'] = 'IMEX'
         # Change time discretization scheme
@@ -186,7 +191,8 @@ class TestTransient:
         # Run
         automated_output_check(capsys, sinusoidal_transient, [1e-4, 2e-3])
 
-    def test_sinusoidal_imex_sbdf_dg(self, capsys: CaptureFixture, sinusoidal_transient: ConfigParser) -> None:
+    def test_sinusoidal_imex_sbdf_dg(self, capsys: CaptureFixture,
[…]
