+++ docs/getting_started/installation_guide.rst
@@ -4,13 +4,12 @@
 Installation Guide
 ==================
 
-To begin using OpenCMP clone the `GitHub repository <https://github.com/uw-comphys/opencmp>`_ and install the dependencies listed below. It is recommended that new users start with the tutorials. Examples of the different models available in OpenCMP can also be found in the "Examples" folder.
+To begin using OpenCMP:
 
-Dependencies
-------------
+1) Clone the `GitHub repository <https://github.com/uw-comphys/opencmp>`_.
+2) Install the following dependencies:
 
-OpenCMP uses Python 3.7+ and requires the following libraries:
-
+* Python 3.7+ 
 * NGSolve version 6.2.2101 or later (`link <https://ngsolve.org/downloads>`_)
 * configparser (:code:`pip install configparser`)
 * pyparsing (:code:`pip install pyparsing`)
@@ -18,5 +17,13 @@ OpenCMP uses Python 3.7+ and requires the following libraries:
 * tabulate (:code:`pip install tabulate`)
 * meshio (:code:`pip install meshio`)
 * pytest (:code:`pip install pytest`)
+
+3) Install the :code:`opencmp` module using :code:`pip3 install .` from the top-level directory. 
+4) Optionally, run the unit tests :code:`python -m pytests` from the top-level directory.
+5) Go through the tutorials and other examples found in the "examples/" folder referring to the `tutorial instructions <https://opencmp.io/tutorials/index.html>`_.
+
+Dependencies
+------------
+
 
 .. warning:: OpenCMP is primarily tested on Linux (Ubuntu 20.04-21.01) and MacOS. The main functionality should work on Windows but users may have issues exporting results to .vtk.
