### 2021-12-15 e4dd1d2
- Converted JOSS manuscript image from SVG to PDF for compilation support - Corrected some spelling from UK to US.
