### 2021-12-15 #3742 @nasserma
I do not understand the issue, it successfully compiles using the preview link,

https://whedon.theoj.org/

using,

https://github.com/uw-comphys/opencmp/

with the branch specified as _publications_
