### 2021-12-16 #3742 @lucydot
Hi @nasserma yes - you are right - it does compile correctly using the preview link you have provided above.

@bonh @WilkAndy I suggest that as a workaround for compiling the paper you use the link [https://whedon.theoj.org/](https://whedon.theoj.org/]) (with `https://github.com/uw-comphys/opencmp/` as the repo and the branch specified as `publications`).

I'll look into why whedon command `@whedon generate pdf from branch publications` isn't working.
