### 2021-12-20 #3742 @lucydot
@openjournals/devs - pinging you here as we are having problems compiling the paper with the command `@whedon generate pdf from branch publications`. Compiling using the link [https://whedon.theoj.org/](https://whedon.theoj.org/) with `https://github.com/uw-comphys/opencmp/` as the repo and `publications` as the branch is working.
