### 2022-01-03 uw-comphys/opencmp#18 @bonh
JOSS: documentation https://opencmp.io/
While reading through your documentation I will add notes for further discussion here.
