### 2022-01-03 uw-comphys/opencmp#18 @bonh
https://opencmp.io/getting_started/installation_guide.html, add references or at least links to the dependencies.
