### 2022-01-03 uw-comphys/opencmp#18 @bonh
> Optionally, run the unit tests `python -m pytests` from the top-level directory.
Perhaps you mean `python -m pytest`?
