### 2022-01-03 uw-comphys/opencmp#18 @bonh
executing `python -m pytest` eats up all my resources. Perhaps you should add a flag or something to limit cores or threads (not sure how)?
