### 2022-01-03 uw-comphys/opencmp#18 @bonh
The test run for a really long time; perhaps you should add a warning to the text or something.
