### 2022-01-03 uw-comphys/opencmp#19 @bonh
JOSS: pytest
A test fails on my machine. Installed all the dependencies via `conda`.

[pytest.txt](https://github.com/uw-comphys/opencmp/files/7801784/pytest.txt)

openjournals/joss-reviews#3742
