### 2022-01-03 uw-comphys/opencmp#19 @Alex-Vasile
That's strange. I have just rerun it and works on my machine.

Which version of NGSolve is installed and what does `conda list` return?
