### 2022-01-03 uw-comphys/opencmp#19 @bonh
Output is [conda.txt](https://github.com/uw-comphys/opencmp/files/7803799/conda.txt) including

> ngsolve                   6.2.2105         py38h3277127_0    ngsolve

Do you see anything suspicious? Let me know, happy to try!
