### 2022-01-05 uw-comphys/opencmp#19 @Alex-Vasile
Okay, something weird is going on. I'm currently thinking it's something to do with Conda. I have tried with system python3.8 and it works. What's curious is that you have ngsolve 6.2.2105. When I installed ngsolve using conda, following their instructions [here](https://ngsolve.org/downloads), I get 6.2.2009.

Did you install ngsolve in conda using their steps? And it looks like you're on linux based on the conda.txt. Is that correct? If so what version?

Could you try installing ngsolve using apt-get and see if you have the same issue?
