### 2022-01-10 #3742 @lucydot
Hello @nasserma @Alex-Vasile and our reviewers @bonh @WilkAndy -

I'm aware this review has been going for > 3months so thought I'd do a quick summary of where we are at now, and try to establish a timeframe for the review.

Looking at the issues that are open:

- [uw-comphys/opencmp#6](https://github.com/uw-comphys/opencmp/issues/6) relates to distributed memory parallelisation (MPI) and there is a question as to whether this is needed to demonstrate substantial scholarly effort, as stated in the JOSS guidelines [here](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). @bonh @WilkAndy Note that substantial scholarly effort is addressed in the closed issue [uw-comphys/opencmp#7](https://github.com/uw-comphys/opencmp/issues/7), so this criteria may be met through other means.
- [uw-comphys/opencmp#8](https://github.com/uw-comphys/opencmp/issues/8) relates to the paper. It looks like we are waiting on an example demonstrating substantial scholarly effort to be included. @nasserma are you waiting on further feedback from the reviewers on this item?
- [uw-comphys/opencmp#18](https://github.com/uw-comphys/opencmp/issues/18). This asks for a small adjustment to the documentation.
- [uw-comphys/opencmp#19](https://github.com/uw-comphys/opencmp/issues/19). There is a failing test which is currently under discussion.

@bonh @WilkAndy, with the above in mind, could I ask you to update your tick-boxes (if appropriate) so I can get a feel for where we are at in the review process?
