### 2022-01-10 uw-comphys/opencmp#19 @bonh
Yes, I followed the instructions. 6.2.2105 should be the right version, see https://anaconda.org/NGSolve/ngsolve. But if you want, I can pull 6.2.2009 and try if it works (in that case, you should add a version warning to your documentation)?

I am on `Ubuntu 20.04.3 LTS` with kernel `5.11.0-36-generic`.

I would rather not use apt to install ngsolve (only as a last resort)...
