### 2022-01-10 #3742 @bonh
Hi,
- uw-comphys/opencmp#6: After I am able to get the software to run on my computer I will check out the 3D simulation case. If it works I consent to close this issue. I am already convinced of the substantial scholarly effort (see the tick-box).
- uw-comphys/opencmp#18: this is ongoing while I walk through the documentation. Right now I am stuck at testing right after installation, see uw-comphys/opencmp#19.
