### 2022-01-10 uw-comphys/opencmp#19 @Alex-Vasile
Please try installing with the apt version, their conda installation option is labelled as experimental.

I shall add a note to the documentation to specify that the conda install not be used until it is not experimental.
