### 2022-01-10 1196e5d
Changes in responce to #18
- Changed pytests to use only use 1 thread for each test
- Added flag to all of the very slow tests so that they don't run by default
- Made changes so that OpenCMP can be used without the optional python dependencies as long their functionality is not being called
- Added support for running pytests in parallel
- Updated documentation
- Moved some python files out of the main folder and into subfolders

In responce to #18.
