### 2022-01-10 uw-comphys/opencmp#18 @Alex-Vasile
The tests now use only a single thread to run. To run the slow tests, the flag `--runslow` must be used. The README.md inside pytests/ has further info about this.

[Documentation](https://opencmp.io/getting_started/installation_guide.html) was also updated.
