### 2022-01-25 #3742 @lucydot
Hi @bonh @WilkAndy, have either of you been able to succesfully run the tests after installation? As highlighted above, this is the current sticking point. The relevant issue thread (uw-comphys/opencmp#19) has been updated and apt-get has been suggested rather than using conda for install.
