### 2022-01-28 uw-comphys/opencmp#18 @bonh
Thank you, that is helpful!

What about that one?

> https://opencmp.io/getting_started/installation_guide.html, add references or at least links to the dependencies.
