### 2022-01-28 uw-comphys/opencmp#18 @bonh
Mhh, no, most likely I am doing something wrong. What is the command to run all the tests at once (including the slow tests)?
