### 2022-01-28 uw-comphys/opencmp#18 @bonh
From the `README` inside `pytests`: `pytest --runslow` but that results in `ERROR: usage: pytest [options] [file_or_dir] [file_or_dir] [...]
pytest: error: unrecognized arguments: --runslow
  inifile: None
  rootdir: /root/opencmp`

I am using a docker with ubuntu and installed everything using apt and pip3.
