### 2022-01-28 uw-comphys/opencmp#18 @bonh
Perhaps it might be best (and more easier for me) if you could provide a docker image with everything one needs to use opencmp preinstalled?
