### 2022-01-28 uw-comphys/opencmp#18 @nasserma
> Thank you, that is helpful!
>
> What about that one?
>
> > https://opencmp.io/getting_started/installation_guide.html, add references or at least links to the dependencies.

We modified this page as well to provide a link to the NGSolve installation page and information required to install the other dependencies using pip. Are there additional changes that you would like us to add?
