### 2022-01-28 uw-comphys/opencmp#18 @nasserma
> `SKIPPED [1] pytests/full_system/ins/test_ins.py:79: need --runslow option to run`, perhaps explain `runslow` in the installation guide https://opencmp.io/getting_started/installation_guide.html?
>
> And I can start the tests only with `python3 -m pytest -runslow`, so only one dash before runslow?

We will take a look at this, the --runslow option does not seem to be consistently implemented (should work using pytest script, but does not work using python -m pytest).
