### 2022-01-28 uw-comphys/opencmp#18 @bonh
> > Thank you, that is helpful!
> > What about that one?
> > > https://opencmp.io/getting_started/installation_guide.html, add references or at least links to the dependencies.
>
> We modified this page as well to provide a link to the NGSolve installation page and information required to install the other dependencies using pip. Are there additional changes that you would like us to add?

I am just suggesting that you add references, for example from that page https://ngsolve.org/documentation/publications. And links to https://github.com/pyparsing/pyparsing, https://github.com/seung-lab/euclidean-distance-transform-3d/. And any other links to important software or references. To give credit, don't you agree?
