### 2022-01-28 uw-comphys/opencmp#18 @Alex-Vasile
> `SKIPPED [1] pytests/full_system/ins/test_ins.py:79: need --runslow option to run`, perhaps explain `runslow` in the installation guide https://opencmp.io/getting_started/installation_guide.html?
>
> And I can start the tests only with `python3 -m pytest -runslow`, so only one dash before runslow?

We have a description of `-runslow` in the README.md inside pytests/, which the installation guide points users towards for more information about running the tests.
