### 2022-02-05 uw-comphys/opencmp#19 @bonh
`123 passed in 926.30s (0:15:26)`

Awesome!:-)
