### 2022-02-08 9ffc2c0
* Reordered examples so that the previous Example 1 is now part of the getting-started documented, shifted all remaining examples down by 1. * Corrected typo in Example 6, now Example 5. * Added Example 10, based on an existing 3D immersed cylinder flow benchmark. * Corrected issues with --runslow option for pytests * Minor refactoring of code. * Fixed an issue with the config parser that caused it to crash when a single boundary was given to the surface traction calculation code. * Changed `sol_to_vtu()` so that by default it will respect the `num_threads` setting in the config file. Added an
[…]
