+++ publications/2108_JOSS/paper.bib
@@ -1,3 +1,14 @@
+@Article{Nguyen2018,
+  author    = {Nguyen, Lam H. and Stoter, Stein K.F. and Ruess, Martin and Sanchez Uribe, Manuel A. and Schillinger, Dominik},
+  journal   = {Int. J. Numer. Methods Eng.},
+  title     = {The diffuse {N}itsche method: Dirichlet constraints on phase-field boundaries},
+  year      = {2018},
+  number    = {4},
+  pages     = {601-633},
+  volume    = {113},
+  doi       = {10.1002/nme.5628},
+}
+
 @Book{Ferziger2002,
   title     = {Computational Methods for Fluid Dynamics},
   publisher = {Springer-Verlag},
@@ -47,7 +58,7 @@
   isbn      = {3-540-66787-3},
 }
 
-@Unpublished{Monte2021,
+@Article{Monte2021,
   title={A diffuse interface method for simulation-based screening of heat transfer processes with complex geometries},
   author={Monte, Elizabeth J and Lowman, James and Abukhdeir, Nasser Mohieddin},
   journal={The Canadian Journal of Chemical Engineering},
@@ -55,6 +66,15 @@
   publisher={Wiley Online Library},
   doi	   = {10.1002/cjce.24320},
 }
+
+@mastersthesis{Monte2021a
+author={{Monte, Elizabeth J}},
+title={OpenCMP: An Open-Source Computational Multiphysics Package},
+year={2021},
+publisher="UWSpace",
+url={http://hdl.handle.net/10012/17239}
+}
+
 
 @Article{Mittal2005,
   author  = {R. Mittal and G. Iaccarino},
@@ -141,4 +161,3 @@
   organization = {ANSYS Inc.},
   url          = {https://www.ansys.com/products/fluids/ansys-fluent},
 }
-
