+++ pytests/full_system/control/test_control.py
@@ -16,7 +16,7 @@
 ########################################################################################################################
 
 from pytest import CaptureFixture, fixture
-from pytests.full_system.helpers import automated_output_check
+from opencmp.helpers.testing import automated_output_check
 from opencmp.config_functions import ConfigParser
 
 # TODO: Figure out how quickly the poisson model should reach SS.
