+++ publications/2108_JOSS/paper.bib
@@ -67,7 +67,7 @@
   doi	   = {10.1002/cjce.24320},
 }
 
-@mastersthesis{Monte2021a
+@mastersthesis{Monte2021a,
 author={{Monte, Elizabeth J}},
 title={OpenCMP: An Open-Source Computational Multiphysics Package},
 year={2021},
