### 2022-02-08 uw-comphys/opencmp#6 @nasserma
Based on your comments here,

https://github.com/openjournals/joss-reviews/issues/3742#issuecomment-1008759727

I am closing this issue, however, we have added MPI functionality to our ongoing projects list.
