### 2022-02-08 uw-comphys/opencmp#18 @nasserma
We have made changes to our documentation to account for issues with the --runslow option using the pytest module and, based on you indicating that you were successfully able to run all tests here,

https://github.com/uw-comphys/opencmp/issues/19#issuecomment-1030589241

I am closing this issue.
