### 2022-02-08 #3742 @nasserma
Hi @lucydot , @bonh was able to run all of our tests, please see the thread here,

https://github.com/uw-comphys/opencmp/issues/19

there was an issue with the pytest module's parsing of the --runslow option, which is corrected in our documentation. From my understanding we have addressed all of the reviewers' comments and, with our most recent version, now have two tutorials which demonstrate 3D simulations both using the diffuse-interface method and flow past an immersed cylinder.

The last open issue is,

https://github.com/uw-comphys/opencmp/issues/8

but we have addressed all of the issues, apart from adding simulation results images to the paper. We can do this, but since we provide links to the examples/tutorials and images/movies are available through our website, we think that is reasonable unless there is a strong objection.
