### 2022-02-10 #3742 @lucydot
Hello @nasserma  - thank you for the update. @bonh @WilkAndy if you are able to give us updates on your review in the next fortnight or so that would very useful. I'm aware this review has been taking some time, hopefully now we are at a stage where the functionality and central claims of the software can be assessed. If you have any questions or are unavailable to review over the immediate time period, please let me know.
