### 2022-02-11 #3742 @WilkAndy
I'm kind of pulling my hair out over here in Australia!!    I'm pretty sure i installed and tested the software, but have since got a new computer and am having lots of trouble installing ngsolve.  It's a real beast with the conda version not working, the mac version not working properly with opencmp, and properly testing opencmp being problematic (for me) on windows.

I'm disappointed that the authors don't want to enhance their paper in various ways i suggested above, because i feel it'd encourage readers to try opencmp, but that's no reason for rejecting the manuscript.   So, i think i "just" have to install ngsolve.
