### 2022-02-11 uw-comphys/opencmp#18 @bonh
What about this? Please engage in a discussion with me - I am trying to help here!

>I am just suggesting that you add references, for example from that page >https://ngsolve.org/documentation/publications. And links to https://github.com/pyparsing/pyparsing, >https://github.com/seung-lab/euclidean-distance-transform-3d/. And any other links to important software or >references. To give credit, don't you agree?
