### 2022-02-11 #3742 @bonh
I get the feeling that you, @nasserma, and your coauthors find me rather annoying to be honest. If so, I am really sorry, but I think it is my duty to be thorough and a little pushy. Sadly, you seldom engage in a discussion with me.

I would like to reopen uw-comphys/opencmp/issues/18 as I am not done with the documentation yet. Same for uw-comphys/opencmp/issues/6 as I was stuck with testing and did not look at the 3D simulations yet. If you close the issues I loose track, sorry about that.

I am continue now with trying out a few simulations but I am convinced by now that this will be publishable.

@WilkAndy I ended up putting it into a docker container (Ubuntu and then a lot of apt get and pip from inside the container). If you want I can try to share it with you (it is not nice...).
