### 2022-02-11 uw-comphys/opencmp#18 @bonh
Running `python3 visualization` from https://opencmp.io/tutorials/tutorial_9.html requires `matplotlib`, right?
