### 2022-02-11 uw-comphys/opencmp#18 @bonh
On https://opencmp.io/tutorials/tutorial_9.html it seems the equations are missing?
