### 2022-02-11 uw-comphys/opencmp#22 @bonh
JOSS paper
I hope I get that right, but you are using the diffuse interface method solely to represent a solid interface, right? In that case, I suggest that you differentiate strongly from the diffuse interface methods used for multiphase flows (most often where both phases are fluid). To avoid confusion. For example, I failed to find the term "diffuse interface method", which you are using in the paper, in the article by (Nguyen et al. 2018).
