### 2022-02-11 #3742 @nasserma
> I'm kind of pulling my hair out over here in Australia!! I'm pretty sure i installed and tested the software, but have since got a new computer and am having lots of trouble installing ngsolve. It's a real beast with the conda version not working, the mac version not working properly with opencmp, and properly testing opencmp being problematic (for me) on windows.
>
> I'm disappointed that the authors don't want to enhance their paper in various ways i suggested above, because i feel it'd encourage readers to try opencmp, but that's no reason for rejecting the manuscript. So, i think i "just" have to install ngsolve.

Sorry that you are having trouble with the installation using Anaconda and MacOS, we will definitely work with the NGSolve developers to try and find out what issues exist and how to fix them. Similarly to you having issues testing on these platforms, we have issues testing on other platforms since we predominantly use some variation of Linux for desktop and high-performance computing.

Regarding your second comment, if you reread my original response from a few days ago I indicated that we would make this change if there were strong objections. This is something we can complete over the weekend and will push a new version of the manuscript for you to review on Monday.
