### 2022-02-11 uw-comphys/opencmp#22 @bonh
I remember that the other reviewer did find the usage of the term "multiphysics" not exactly suitable. I cannot find your discussion about that point right now (can you show me?). I guess it is somewhat an issue of how to define that term but do you agree with the definition provided on the wikipedia page https://en.wikipedia.org/wiki/Multiphysics_simulation?
