### 2022-02-11 #3742 @bonh
I started to carefully read the paper. Comments are in https://github.com/uw-comphys/opencmp/issues/22.
