### 2022-02-11 #3742 @nasserma
> I get the feeling that you, @nasserma, and your coauthors find me rather annoying to be honest. If so, I am really sorry, but I think it is my duty to be thorough and a little pushy. Sadly, you seldom engage in a discussion with me.
>
> I would like to reopen [uw-comphys/opencmp/issues/18](https://github.com/uw-comphys/opencmp/issues/18) as I am not done with the documentation yet. Same for [uw-comphys/opencmp/issues/6](https://github.com/uw-comphys/opencmp/issues/6) as I was stuck with testing and did not look at the 3D simulations yet. If you close the issues I loose track, sorry about that.
>
> I am continue now with trying out a few simulations but I am convinced by now that this will be publishable.

I am sorry if you get the feeling that I am annoyed with you or am not engaging in conversation. You have to understand, this is a very different peer review process from what I am accustomed to. I am doing my best to adapt to it, but I also must consult with my co-authors before responding, making changes to the code, and revising the manuscript. This takes time, as I am sure you understand.

Regarding the two issues with the code that you found, we are all very sorry for not catching those and will correct this in the future by implement automated testing of the examples, in addition to the unit tests/pytests. These were introduced by some refactoring and it was my responsibility to test the examples. We will add this to our to-do list for the weekend!

Please check back on Monday, and make sure to use the _publications_ branch, which we'll merge into the main branch once all of these issues are resolved.
