### 2022-02-11 uw-comphys/opencmp#22 @nasserma
> I hope I get that right, but you are using the diffuse interface method solely to represent a solid interface, right? In that case, I suggest that you differentiate strongly from the diffuse interface methods used for multiphase flows (most often where both phases are fluid). To avoid confusion. For example, I failed to find the term "diffuse interface method", which you are using in the paper, in the article by (Nguyen et al. 2018).

This is a pretty complicated topic to address here, but we did address it in [Elizabeth's paper](https://doi.org/10.1002/cjce.24320) quite well. There is a great deal of variation of the use of the term diffuse interface in the literature. What we are doing is similar to what is done to simulation segregated flows when the indicator field is meant to approximate the interface from a numerical perspective, so a continuous immersed boundary method, versus approximating the interface from a physical perspective, such as a continuous phase field model (_e.g._ Cahn-Hilliard). From my understanding, there is a significant conceptual difference between immersed boundary _numerical methods_ and phase field _models_. The former are typically used to approximate a sharp-interface, which is possibly evolving and deforming, while the latter is a physical model which must conserve mass, momentum, energy and include interfacial energy.
