### 2022-02-11 uw-comphys/opencmp#22 @bonh
In my opinion the statement of need is very long. The "first barrier" you are observing is present, I generally agree, but your are not completely solving it. So I suggest that instead of using very general and broad statements like "MOOSE [...] and FEniCS [...] have very steep learning curves and are inaccessible to the general scientific and engineering communities.", "One such barrier is a lack of suitable software accessible to the general scientific and engineering community." or "the predominance of the finite volume method for fluid dynamics simulations inherently limits simulation accuracy for a given mesh compared to the finite element method" you pinpoint what your software really wants to improve. With that paragraph you have the whole community crying "not true at all, Sir!" :-D.
