### 2022-02-11 uw-comphys/opencmp#22 @bonh
> There is a great deal of variation of the use of the term diffuse interface in the literature.

I totally agree! I just think that if you are able to define it more sharply (pun intended) your paper improves. Perhaps use a footnote or something?
