### 2022-02-11 uw-comphys/opencmp#22 @nasserma
> I remember that the other reviewer did find the usage of the term "multiphysics" not exactly suitable. I cannot find your discussion about that point right now (can you show me?). I guess it is somewhat an issue of how to define that term but do you agree with the definition provided on the wikipedia page https://en.wikipedia.org/wiki/Multiphysics_simulation?

Regarding the Wikipedia definition, I think it provides a reasonable definition, for example from it "simultaneous simulation of the physical stress on an object and the temperature distribution of the object would be considered a multiphysics simulation". This describes a simulation involving conservation of momentum and conservation of energy.

Regarding the discussion with the other reviewer, please see here,

https://github.com/openjournals/joss-reviews/issues/3742#issuecomment-937759257
