### 2022-02-11 uw-comphys/opencmp#22 @bonh
I do remember a discussion about adding a figure and a brief simulation study to the paper, right? I think that would be really valuable and I suggest you use the diffuse interface test case, perhaps together with NS and heat transfer? Would that be possible?
