### 2022-02-11 uw-comphys/opencmp#22 @bonh
> This describes a simulation involving conservation of momentum and conservation of energy.

So is tutorial 9 both of them? I am not really aware of that yet...
