### 2022-02-11 uw-comphys/opencmp#22 @nasserma
I am going to add a section over the weekend which provides high-level summarizes of the key examples, linking to their detailed summaries on our website. I will not have anymore time until later today, but I am going to try to summarize all of the revisions from this thread that I will work on over the weekend. If you could confirm that I get things right tonight or tomorrow AM, I would appreciate it. That way Monday's revision is as close to what is desired as feasible.
