### 2022-02-11 uw-comphys/opencmp#23 @Alex-Vasile
I'm kind of pulling my hair out over here in Australia!!    I'm pretty sure i installed and tested the software, but have since got a new computer and am having lots of trouble installing ngsolve.  It's a real beast with the conda version not working, the mac version not working properly with opencmp, and properly testing opencmp being problematic (for me) on windows.
__Originally posted by @WilkAndy in https://github.com/openjournals/joss-reviews/issues/3742#issuecomment-1035697198__
