### 2022-02-11 uw-comphys/opencmp#23 @Alex-Vasile
Hi, if you can provide me with some more info on the troubles you're having with the mac version, I can try to help you sort it out
