### 2022-02-11 uw-comphys/opencmp#23 @nasserma
One of us has it working on MacOS, so we should be able to help.
