### 2022-02-15 uw-comphys/opencmp#22 @nasserma
I am running behind on this, will try to finish the revision tomorrow and will push to the _publication_ branch.
