### 2022-02-17 uw-comphys/opencmp#22 @nasserma
> In my opinion the statement of need is very long. The "first barrier" you are observing is present, I generally agree, but your are not completely solving it. So I suggest that instead of using very general and broad statements like "MOOSE [...] and FEniCS [...] have very steep learning curves and are inaccessible to the general scientific and engineering communities.", "One such barrier is a lack of suitable software accessible to the general scientific and engineering community." or "the predominance of the finite volume method for fluid dynamics simulations inherently limits simulation accuracy for a given mesh compared to the finite element method" you pinpoint what your software really wants to improve. With that paragraph you have the whole community crying "not true at all, Sir!" :-D.

So i am still working on the revised paper, but I think we have addressed all of the concerns from this thread. Given that there is a controversial paragraph in the Statement of Need section, I want to get some feedback before pushing all of the changes. This could save some back and forth,

_However, there are several barriers to more wide spread use of simulations. One such barrier is the lack of user-friendly finite element-based open-source simulation software. Many of the most widely-used computational multiphysics software packages, such as COMSOL Multiphysics [@comsol] or ANSYS Fluent [@fluent], are closed-source and cost-prohibitive. Furthermore, the predominance of the finite volume method for fluid dynamics simulations inherently limits simulation accuracy for a given mesh compared to the finite element method [@Ferziger2002]. Many high-quality open-source packages exist which use the finite element method, such as the computational multiphysics software MOOSE [@Permann2020] or the finite element libraries NGSolve [@ngsolve] and FEniCS [@Alnaes2015]. However, they require that the user has a relatively detailed understanding of the finite element method, such as derivation of the weak formulation of the problem, along with programming background. Alternatively, open-source finite volume-based packages such [OpenFOAM](https://openfoam.org/) and [SU2](https://su2code.github.io/) are more accessible to the broader scientific and engineering community in that they require a less detailed understanding of numerical methods and programming, instead requiring an understanding of the command line interface (CLI) and configuration through a set of configuration files._
