### 2022-02-17 uw-comphys/opencmp#22 @nasserma
I realize that some of the statements might still be controversial, but I really think there is a big difference between the accessibility and user-friendliness of software designed to be libraries/frameworks (MOOSE, FEnICS, NGSolve, _et al_) versus those which have a CLI-based user interface (OpenFOAM, SU2, _et al_). Many scientists and engineers do not have the programming and numerical methods background to use the former, and tend to use the latter.

Separately there are a few GUI-based open-source multiphysics packages, such as Elmer, but I have not found them as conducive to large-scale research simulations as the ones previously listed.
