### 2022-02-22 8edb7ff
- Revised JOSS manuscript to include short descriptions and simulation images from tutorials 8-10, see Issue #8 - Revised JOSS manuscript to address comments from reviewers, see Issue #22 - Modified tutorial 8 to have a cosine ramp of inlet velocity to remove numerical instability from using no flow initial condition.
