+++ publications/2108_JOSS/paper.bib
@@ -1,3 +1,15 @@
+
+@article{bayraktar2012benchmark,
+  title={Benchmark computations of 3D laminar flow around a cylinder with CFX, OpenFOAM and FeatFlow},
+  author={Bayraktar, Evren and Mierka, Otto and Turek, Stefan},
+  journal={International Journal of Computational Science and Engineering},
+  volume={7},
+  number={3},
+  pages={253--266},
+  year={2012},
+  publisher={Inderscience Publishers Ltd}
+}
+
 @Article{Nguyen2018,
   author    = {Nguyen, Lam H. and Stoter, Stein K.F. and Ruess, Martin and Sanchez Uribe, Manuel A. and Schillinger, Dominik},
   journal   = {Int. J. Numer. Methods Eng.},
@@ -67,7 +79,7 @@
   doi	   = {10.1002/cjce.24320},
 }
 
-@mastersthesis{Monte2021a,
+@mastersthesis{Monte2021a
 author={{Monte, Elizabeth J}},
 title={OpenCMP: An Open-Source Computational Multiphysics Package},
 year={2021},
