+++ publications/2108_JOSS/paper.bib
@@ -1,5 +1,5 @@
 
-@article{bayraktar2012benchmark,
+@article{Bayraktar2012,
   title={Benchmark computations of 3D laminar flow around a cylinder with CFX, OpenFOAM and FeatFlow},
   author={Bayraktar, Evren and Mierka, Otto and Turek, Stefan},
   journal={International Journal of Computational Science and Engineering},
@@ -79,7 +79,7 @@
   doi	   = {10.1002/cjce.24320},
 }
 
-@mastersthesis{Monte2021a
+@mastersthesis{Monte2021a,
 author={{Monte, Elizabeth J}},
 title={OpenCMP: An Open-Source Computational Multiphysics Package},
 year={2021},
