### 2022-02-22 834c82f
Corrected filename issues with paper.md and images/
