+++ publications/2108_JOSS/paper.md
@@ -102,29 +102,29 @@ Several examples of usage of OpenCMP are present in ref. [@Monte2021a] and also 
 
 [Tutorial 8](https://opencmp.io/tutorials/tutorial_8.html) demonstrates the usage of OpenCMP to solve a multiphysics problem, two-dimensional transient multi-component incompressible flow around a set of immersed circular objects within a rectangular channel, shown below.
 
-![](images/tutorial8_geom.png)
+![](images/tutorial_8_geom.png)
 
 The mixture is composed of two components (A, B) where A undergoes an irreversible reaction A->B. A parabolic inlet flow of pure A is imposed, such that the mixture undergoes convection, reaction, and molecular diffusion throughout the channel. Sample simulation results of the steady-state are shown resulting from a cosine ramp of the inlet velocity an initial condition with no flow and pure A.
 
-![](images/tutorial8_results_3s_a.png)
+![](images/tutorial_8_results_3s_a.png)
 
-![](images/tutorial8_results_3s_b.png)
+![](images/tutorial_8_results_3s_b.png)
 
 [Tutorial 9](https://opencmp.io/tutorials/tutorial_9.html) demonstrates how to use the diffuse interface method to approximate complex geometries with nonconformal structured quadrilateral/hexahedral meshes, versus the use of a traditional unstructured mesh which conforms to the complex geometry boundary. The sample problem is based on simulation of heat transfer within an LED heat sink from ref. [@Monte2021], with a geometry shown below.
 
-![](images/tutorial9_geometry.png)
+![](images/tutorial_9_geometry.png)
 
 The base of the heat sink is exposed to a spatially varying heat flux profile, corresponding to heat generated by an LED assembly, with convective heat transfer conditions assumed on the exterior fins. A script is provided for ease of post-processing visualization, showing the diffuse-interface boundaries (comparable to the geometry above) and steady-state temperature profile.
 
-![](images/tutorial9_cross_section.png)
+![](images/tutorial_9_cross_section.png)
 
 [Tutorial 10](https://opencmp.io/tutorials/tutorial_10.html) demonstrate using OpenCMP to perform a standard benchmark for 3D flow around an immersed cylinder [@Bayraktar2012] under laminar flow conditions. Both continuous and discontinuous Galerkin finite element solver configurations are provided, defaulting to the continuous Galerkin variation due to reduced memory constraints. The geometry is shown below with parabolic inlet velocity resulting in Re=20.
 
-![](images/tutorial10_geometry.png)
+![](images/tutorial_10_geometry.png)
 
 This example also uses the built-in error analysis functionality of OpenCMP to automatically compute the force vector on the immersed cylinder through integration of the surface traction over its boundary. This facilitates the calculation of the resulting drag and lift forces on the immersed cylinder. A visualization of the three-dimensional steady-state velocity field is shown below.
 
-![](images/tutorial10_result.png)
+![](images/tutorial_10_result.png)
 
 Several additional examples of usage of OpenCMP in tutorial form are available via the [website](https://opencmp.io/tutorials/).
 
