### 2022-02-22 a2dc0e0
Another change to paper.md to improve the layout.
