+++ publications/2108_JOSS/paper.md
@@ -112,11 +112,9 @@ The mixture is composed of two components (A, B) where A undergoes an irreversib
 
 [Tutorial 9](https://opencmp.io/tutorials/tutorial_9.html) demonstrates how to use the diffuse interface method to approximate complex geometries with nonconformal structured quadrilateral/hexahedral meshes, versus the use of a traditional unstructured mesh which conforms to the complex geometry boundary. The sample problem is based on simulation of heat transfer within an LED heat sink from ref. [@Monte2021], with a geometry shown below.
 
-![](images/tutorial_9_geometry.png)
-
 The base of the heat sink is exposed to a spatially varying heat flux profile, corresponding to heat generated by an LED assembly, with convective heat transfer conditions assumed on the exterior fins. A script is provided for ease of post-processing visualization, showing the diffuse-interface boundaries (comparable to the geometry above) and steady-state temperature profile.
 
-![](images/tutorial_9_cross_section.png)
+![](images/tutorial_9_geometry.png) ![](images/tutorial_9_cross_section.png)
 
 [Tutorial 10](https://opencmp.io/tutorials/tutorial_10.html) demonstrate using OpenCMP to perform a standard benchmark for 3D flow around an immersed cylinder [@Bayraktar2012] under laminar flow conditions. Both continuous and discontinuous Galerkin finite element solver configurations are provided, defaulting to the continuous Galerkin variation due to reduced memory constraints. The geometry is shown below with parabolic inlet velocity resulting in Re=20.
 
