### 2022-02-22 #3742 @nasserma
@lucydot @WilkAndy @bonh I have pushed a revised manuscript with changes meant to address all of the issues and comments from the open issues (#8, 22, 23). I am having trouble with the layout of images, but I want to get an otherwise complete revision out while I sort out using the JOSS docker image to compile locally.

Please let me know what you all think!
