### 2022-02-27 uw-comphys/opencmp#22 @bonh
> This is a pretty complicated topic to address here, but we did address it in [Elizabeth's paper](https://doi.org/10.1002/cjce.24320) quite well. There is a great deal of variation of the use of the term diffuse interface in the literature. What we are doing is similar to what is done to simulation segregated flows when the indicator field is meant to approximate the interface from a numerical perspective, so a continuous immersed boundary method, versus approximating the interface from a physical perspective, such as a continuous phase field model (e.g. Cahn-Hilliard). From my understanding, there is a significant conceptual difference between immersed boundary numerical methods and phase field models. The former are typically used to approximate a sharp-interface, which is possibly evolving and deforming, while the latter are physical models which must conserve mass, momentum, energy and include interfacial energy.

That's a really interesting viewpoint and I think I am getting what you aiming for. I agree that there is a conceptual difference, however, I cannot really spot a mathematical difference between phase field models and diffuse interface method as immersed boundary method. I guess, that if the same terms are used for different things in slightly different fields of research you might want to add a footnote to clarify your vocabulary. Otherwise, as I said above, I think, reader might have the wrong expectation about your code.
