### 2022-02-27 uw-comphys/opencmp#22 @bonh
> I do remember a discussion about adding a figure and a brief simulation study to the paper, right? I think that would be really valuable and I suggest you use the diffuse interface test case, perhaps together with NS and heat transfer? Would that be possible?

> Please see tutorial 8 for the convection/reaction/diffusion example, I rearranged the tutorials with the last revision.

I like tutorial 8! But you are not using your diffuse interface method to represent the "pillars", right? Why not? I think such an example would be really great for your code, i.e., multiphysics plus immersed boundary method.
