### 2022-02-27 uw-comphys/opencmp#22 @bonh
> So i am still working on the revised paper, but I think we have addressed all of the concerns from this thread. Given that there is a controversial paragraph in the Statement of Need section, I want to get some feedback before pushing all of the changes. This could save some back and forth,

Sorry, I did not get that, now it's too late I will comment on what you wrote in the paper, right?
