### 2022-02-27 uw-comphys/opencmp#19 @bonh
Sorry to reopen that one, but with the new revision I am again not able to run all the tests. This is what I did:

1. Add the top level folder `opencmp` as it is cloned from the branch `main` to `PYTHONPATH.
2. Run `pytest --runslow` from `opencmp`

It returns the message:
```
ERROR: usage: pytest [options] [file_or_dir] [file_or_dir] [...]
pytest: error: unrecognized arguments: --runslow
  inifile: None
  rootdir: /opencmp
```

I think I am missing something here...
