### 2022-02-27 uw-comphys/opencmp#22 @bonh
It just came to my mind, that accessibility is really important to you, so why not add it to the title of your paper? Just an idea, do not really have anything else to add.
