### 2022-02-27 uw-comphys/opencmp#22 @bonh
> I do remember a discussion about adding a figure and a brief simulation study to the paper, right? I think that would be really valuable[...].

Did I miss something here? Why do you not add a color figure and a tutorial case to your paper? I think the editor agreed somewhere that you do not have to be strictly below the word count limit.
