### 2022-02-27 uw-comphys/opencmp#22 @bonh
Please, please comment on my second post from 16 days ago. I still think that my questions are very valid!
