### 2022-02-27 #3742 @bonh
I read through the paper again and added some new discussion points in https://github.com/uw-comphys/opencmp/issues/22. And there are some old points which have not been addressed yet.
