### 2022-02-28 uw-comphys/opencmp#19 @Alex-Vasile
We've been having similar issues within our group when running the tests with `pytest --runslow`. Though it is not consistent. On seemingly identical setups it works for some but not for others. We believe it's a pytest issue.

If you rest the tests as `pytest --runslow pytests/` (i.e. provide a path or a file to it) then pytests seems to work. The README inside the pytests folder is update to reflect this.
