### 2022-03-02 uw-comphys/opencmp#23 @nasserma
@WilkAndy can you confirm that you are still having installation issues on MacOS? Otherwise we can close this issue,

https://github.com/uw-comphys/opencmp/issues/23
