### 2022-03-02 uw-comphys/opencmp#22 @nasserma
> Please, please comment on my second post from 16 days ago. I still think that my questions are very valid!

I'm sorry, I though I responded to everything prior to your new posts from a few days again, could you explicitly mention which comment that you are referring to?

> I guess, that if the same terms are used for different things in slightly different fields of research you might want to add a footnote to clarify your vocabulary. Otherwise, as I said above, I think, reader might have the wrong expectation about your code.

I will add a statement that refers to two reviews which provide background on different the terminology,

https://doi.org/10.1146/annurev.fluid.37.061903.17574
https://doi.org/10.1146/annurev.fluid.30.1.139

Again, there is no clear definitive usage of this terminology across disciplines and sub-disciplines, but I think providing these references will provide users with some additional clarity.

> I like tutorial 8! But you are not using your diffuse interface method to represent the "pillars", right? Why not? I think such an example would be really great for your code, i.e., multiphysics plus immersed boundary method.

As @Alex-Vasile mentioned, we are still working on implementing DI method for the multicomponent Navier-Stokes solver, but for Tutorial 8 we would need to change it to a 3D problem to see any real benefit from the DI method.

> Did I miss something here? Why do you not add a color figure and a tutorial case to your paper? I think the editor agreed somewhere that you do not have to be strictly below the word count limit.

We did, please see [the updated version of the manuscript](https://raw.githubusercontent.com/openjournals/joss-papers/joss.03742/joss.03742/10.21105.joss.03742.pdf).

> It just came to my mind, that accessibility is really important to you, so why not add it to the title of your paper? Just an idea, do not really have anything else to add.

Accessibility means something very specific at our university (and in Canada), so I would not be comfortable using it in the title of our paper since there is no way to specify exactly what we mean. We will add a statement or two in the manuscript regarding this, though, hopefully that will address your comment.
