### 2022-03-02 uw-comphys/opencmp#22 @nasserma
Here I am going through your Devil's Advocate comment above sequentially:

> > One such barrier is a lack of suitable software accessible to the general scientific and engineering community.
>
> What do you mean by suitable? It follows next, I guess (closed-source, cost, finite volume, steep learning curve).

Exactly, the following statements clarify and expand on this initial claim.

> > Many of the most widely-used computationathousendsl multiphysics software packages, such as COMSOL Multiphysics (COMSOL Multiphysics ®, n.d.) or ANSYS Fluent (Ansys ® 40 Fluent, n.d.), are closed-source and cost-prohibitive.
>
> Why does closed-source mean that a software is not "suitable" to the general scientific and engineering community? If you look for papers and industry projects which use comsol and fluent you will find many, many successful projects, right? Secondly, in my experience, cost of software is not a strong argument for industry (missing support is however!). So what do you mean with "general scientific and engineering community"? Do you exclude industry? thousends

It depends on what you define as a "successful" project, but if we are performing research, closed-source software involves software implementations that are not accessible and, therefore, not reproducible. While there are plenty of research publications in which closed-source software is used, clearly there is an issue when research methods are used that are not documented and verified.

Regarding cost, I have over 10 years of experience with industry both through academic-industry partnerships and engineering consulting. I cannot agree with you at all on this point, based on my own experience and what has been communicated by many of my colleagues and collaborators.

> > Furthermore, the predominance of the finite volume method for fluid dynamics simulations inherently limits simulation accuracy for a given mesh compared to the finite element method (Ferziger & Perić, 2002).
>
> That is a very strong statement! Is that really true? As far as I know, the finite volume method is somewhat similar to the Discontinuous Galerkin method which is somewhat similar to the finite volume method. What page are you referring to in the book by Ferziger? If your statement is true, there have to be advantages of the finite volume method (mass conservation, for example?) against fem, otherwise, a whole bunch of mathematical research groups are wasting there time and grants?! Seems unlikely...

The finite-volume method is, essentially, a zeroth order discontinuous Galerkin method, so the statement is strictly true since DG enables the use of higher-order elements. If you look in the applied math literature, there is a great deal of work on DG and hybrid DG methods. I lent my copy of Ferziger to one of my students, but I will respond shortly with the exact section along with many other references to support this claim after consulting with my colleague in Applied Math whose research is focused on hybrid DG.

> > Open-source packages that uses the finite element method, such as the computational multiphysics software MOOSE (Permann et al., 2020) or the finite element libraries NGSolve (Schöberl, 2020) and FEniCS (Alnaes et al., 2015), have very steep learning curves and are inaccessible to the general scientific and engineering communities.
>
> This is a very subjective statement. Why are the learning curves steep? And calling a software inaccessible is almost insulting, and most likely untrue, because these software packages have thousands of users and successful projects.

We tried to be clear in the revision, the need to know a programming language itself can easily be considered a steep learning curve. OpenCMP requires understanding of the CLI, editing configuration files, etc, but no need to know Python or have a detailed understanding of the numerical methods "under the hood". There is no insult intended and no claims that these different packages are problematic, in fact, we use one of them extensively in OpenCMP.
