### 2022-03-02 #3742 @nasserma
I reopened them.

> > I would like to reopen [uw-comphys/opencmp#18](https://github.com/uw-comphys/opencmp/issues/18) as I am not done with the documentation yet. Same for [uw-comphys/opencmp#6](https://github.com/uw-comphys/opencmp/issues/6) as I was stuck with testing and did not look at the 3D simulations yet. If you close the issues I loose track, sorry about that.
>
> Perhaps you could reopen them?
