### 2022-03-02 #3742 @nasserma
I responded to all of your new comments, however, you indicated that a previous comment was not responded to but I could not determine which one you were referring to.

> I read through the paper again and added some new discussion points in [uw-comphys/opencmp#22](https://github.com/uw-comphys/opencmp/issues/22). And there are some old points which have not been addressed yet.
