### 2022-03-02 uw-comphys/opencmp#22 @bonh
I am talking about that one:

> I find that statement difficult to accept (which does not mean that I am correct here!). The diffuse interface method is to my best knowledge a very broad concept describing a lot of possible equations and implementations. I guess, often the Cahn-Hilliard or Allen-Cahn equations are referenced that way. However, to proof your statement wrong I just have to find one single, public available (what do you mean by public, for example?) implementation (which includes some github code snippets?), even with low quality?

> https://mooseframework.inl.gov/source/kernels/CahnHilliard.html, https://fenicsproject.org/olddocs/dolfin/1.3.0/python/demo/documented/cahn-hilliard/python/documentation.html, and a snippet https://github.com/urbainvaes/cahn-hilliard (which I have not tested, it is four years old! I am just suggesting, that merely implementing some diffuse interface method is trivial!)

> I asked about your diffuse interface method before in https://github.com/uw-comphys/opencmp/issues/7: "You mention that there are no other implementations of the diffuse interface method. Is that really correct? How about the Phase Field implementation of comsol? Can that not be used to simulate "hard" boundaries by choosing super high values for density and viscosity or so?" and you answered: "To our knowledge, this is the only open-source implementation of the diffuse interface method with stabilization for DiricYou continue with: "What you are describing with Comsol's phase field model would not straightforward to implement, would not be stable when using Dirichlet BCs, and is not open-source." I am still not getting the different between your approach of the diffuse interface method and just increasing the density and viscosity in Comsol. Do you suggest that the stability of the numerical procedure used by comsol is insufficient for such a problem?hlet boundary conditions (see Nguyen(2018))."

> By stabilization for Dirichlet boundary conditions do you mean the weak implementation of strong bcs using Nitsche's method?

> You continue with: "What you are describing with Comsol's phase field model would not straightforward to implement, would not be stable when using Dirichlet BCs, and is not open-source." I am still not getting the different between your approach of the diffuse interface method and just increasing the density and viscosity in Comsol. Do you suggest that the stability of the numerical procedure used by comsol is insufficient for such a problem?
