### 2022-03-02 uw-comphys/opencmp#22 @bonh
> As @Alex-Vasile mentioned, we are still working on implementing DI method for the multicomponent Navier-Stokes solver, but for Tutorial 8 we would need to change it to a 3D problem to see any real benefit from the DI method.

So do you want to add it before the paper gets published?
