### 2022-03-02 uw-comphys/opencmp#22 @bonh
Mhhh, I really do not see a color figure or a description of any tutorial case in the new revision of the paper. I am confused...

Can you upload a pdf? To exclude that the compilation is wrong.
