### 2022-03-02 uw-comphys/opencmp#22 @bonh
The remainder of you answer is very satisfying and helpful, thank you!
