### 2022-03-02 uw-comphys/opencmp#18 @nasserma
We added the lines for installing these packages using PIP in the installation guide,

https://opencmp.io/getting_started/installation_guide.html

so with the unique identifier of each package the user could get all info from "pip show IDENTIFIER". If you want us to also include links directory to their websites, I'll do that shortly, however I now realize that it would be better to determine if any of these packages have published papers associated with them and to cite them in the JOSS submission.
