### 2022-03-02 uw-comphys/opencmp#22 @nasserma
> Mhhh, I really do not see a color figure or a description of any tutorial case in the new revision of the paper. I am confused...
>
> Can you upload a pdf? To exclude that the compilation is wrong.

I am super-confused, I get an updated PDF when I use the JOSS Paper Preview link https://whedon.theoj.org/

Please give me an hour or so to determine what is wrong.
