### 2022-03-02 uw-comphys/opencmp#19 @bonh
This is with `pytest --runslow pytests/`: `123 passed in 581.61s (0:09:41)`
and with `pytest pytests/`: `116 passed, 7 skipped in 213.73s (0:03:33)`.

Awesome!
