### 2022-03-02 uw-comphys/opencmp#6 @bonh
I am trying to get tutorial 10 to run. I am on branch `publications`. Following the documentation I try `python3 -m opencmp config_IC` but it fails with
```
importing NGSolve-6.2.2105
Using the default value of False for TRANSIENT, transient.
WARNING: Physical groups detected - Be sure to define them for every geometrical entity.
 Curve elements, order = 2
WARNING: The direct linear solver does not respect the num_threads parameter you may have set.This is an NGSolve issue
Traceback (most recent call last):
  File "/usr/lib/python3.8/runpy.py", line 194, in _run_module_as_main
    return _run_code(code, main_globals, None,
  File "/usr/lib/python3.8/runpy.py", line 87, in _run_code
    exec(code, run_globals)
  File "/root/opencmp/opencmp/__main__.py", line 29, in <module>
    run(config_file_path)
  File "/root/opencmp/opencmp/run.py", line 66, in run
    calc_error(config_parser, solver.model, sol)
  File "/root/opencmp/opencmp/helpers/error.py", line 305, in calc_error
    surface_traction = _surface_traction(sol, model, marker)
  File "/root/opencmp/opencmp/helpers/error.py", line 222, in _surface_traction
    raise ValueError('{} is not a mesh surface.'.format(marker))
ValueError: c is not a mesh surface.
```
