### 2022-03-02 uw-comphys/opencmp#18 @bonh
Well, I was merely suggesting giving credit, yes. Sure, I can always try to find everything myself but I appreciate if developers explicitly acknowledge other peoples work. So in your documentation, I think it would be beneficial to add a link to the github repos, for example. Citing published papers in you own paper is even better.
