### 2022-03-02 #3742 @danielskatz
@xuanxu - Can you please look at this error
> /home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.1.0/gems/theoj-1.2.1/lib/theoj/paper.rb:155:in `failure': Couldn't check the bibtex because branch name is incorrect: publications (Theoj::Error)
