### 2022-03-07 88af202
Added missing DOIs to JOSS bib file.
