+++ publications/2108_JOSS/paper.bib
@@ -7,7 +7,8 @@
   number={3},
   pages={253--266},
   year={2012},
-  publisher={Inderscience Publishers Ltd}
+  publisher={Inderscience Publishers Ltd},
+  doi       = {10.1504/ijcse.2012.048245}
 }
 
 @Article{Nguyen2018,
@@ -57,6 +58,7 @@
   type        = {techreport},
   number      = {LA-UR-03-1560},
   address     = {Los Alamos, NM, USA},
+  doi      = {10.1016/b978-012387582-2/50038-1},
 }
 
 @InBook{Cockburn2000,
@@ -134,6 +136,7 @@ url={http://hdl.handle.net/10012/17239}
   number   = {246},
   pages    = {569--593},
   month    = may,
+  doi      = {10.1090/s0025-5718-03-01552-7},
 }
 
 @Article{Ascher1995,
@@ -155,6 +158,7 @@ url={http://hdl.handle.net/10012/17239}
   year      = {2015},
   address   = {Fitzwilliam College, Cambridge University, UK},
   month     = jul,
+  doi      = {10.1109/iccse.2015.7250335},
 }
 
 @Misc{comsol,
