### 2022-03-09 uw-comphys/opencmp#6 @bonh
Okay, the example runs through, all good. But the visualization of `output/transient.pvd` with paraview v.5.10.0 fails with the errors below. What kind of version of paraview are you using? Or do you recommend a different software?
```
ERROR: In /builds/gitlab-kitware-sciviz-ci/build/superbuild/paraview/src/VTK/IO/XMLParser/vtkXMLParser.cxx, line 379
vtkXMLDataParser (0x18d30190): Error parsing XML in stream at line 8, column 10, byte index 241: junk after document element

ERROR: In /builds/gitlab-kitware-sciviz-ci/build/superbuild/paraview/src/VTK/IO/XML/vtkXMLReader.cxx, line 521
vtkPVDReader (0x17763f80): Error parsing input file.  ReadXMLInformation aborting.

ERROR: In /builds/gitlab-kitware-sciviz-ci/build/superbuild/paraview/src/VTKExtensions/IOCore/vtkPVDReader.cxx, line 133
vtkPVDReader (0x17763f80): Could not read file information

ERROR: In /builds/gitlab-kitware-sciviz-ci/build/superbuild/paraview/src/VTK/Common/ExecutionModel/vtkExecutive.cxx, line 753
vtkPVCompositeDataPipeline (0x18d5d930): Algorithm vtkPVDReader(0x17763f80) returned failure for request: vtkInformation (0x170c33b0)
  Debug: Off
  Modified Time: 386992
  Reference Count: 1
  Registered Events: (none)
  Request: REQUEST_DATA_OBJECT
  ALGORITHM_AFTER_FORWARD: 1
  FORWARD_DIRECTION: 0

ERROR: In /builds/gitlab-kitware-sciviz-ci/build/superbuild/paraview/src/VTKExtensions/IOCore/vtkPVDReader.cxx, line 133
vtkPVDReader (0x17763f80): Could not read file information

ERROR: In /builds/gitlab-kitware-sciviz-ci/build/superbuild/paraview/src/VTK/Common/ExecutionModel/vtkExecutive.cxx, line 753
vtkPVCompositeDataPipeline (0x18d5d930): Algorithm vtkPVDReader(0x17763f80) returned failure for request: vtkInformation (0x170c33b0)
  Debug: Off
  Modified Time: 386992
  Reference Count: 1
  Registered Events: (none)
  Request: REQUEST_DATA_OBJECT
  ALGORITHM_AFTER_FORWARD: 1
  FORWARD_DIRECTION: 0

ERROR: In /builds/gitlab-kitware-sciviz-ci/build/superbuild/paraview/src/VTKExtensions/IOCore/vtkPVDReader.cxx, line 133
vtkPVDReader (0x17763f80): Could not read file information

ERROR: In /builds/gitlab-kitware-sciviz-ci/build/superbuild/paraview/src/VTK/Common/ExecutionModel/vtkExecutive.cxx, line 753
vtkPVCompositeDataPipeline (0x18d5d930): Algorithm vtkPVDReader(0x17763f80) returned failure for request: vtkInformation (0x170c33b0)
  Debug: Off
  Modified Time: 386992
  Reference Count: 1
  Registered Events: (none)
  Request: REQUEST_DATA_OBJECT
  ALGORITHM_AFTER_FORWARD: 1
  FORWARD_DIRECTION: 0
```
