### 2022-03-09 uw-comphys/opencmp#22 @bonh
The only thing open about the paper for me now is https://github.com/uw-comphys/opencmp/issues/22#issuecomment-1056968759. I am really interested in your thoughts!
