### 2022-03-09 uw-comphys/opencmp#22 @nasserma
We removed this statement in the revised manuscript based on your concerns. While, in the context of what we mean by the "diffuse interface method" I think our statement is correct, your point that this can be interpreted much more widely is definitely a point of confusion that would require readers (unreasonably) to both completely understand and agree with our terminology.

_The revised statement now reads "Finally, OpenCMP provides a publicly available implementation of the diffuse interface method with support for stabilized Dirichlet boundary conditions (Nguyen et al., 2018)."_

Later in your comment, you talk about using Comsol's phase field model to capture a complex fluid/solid interface through using a very large viscosity for the solid phase. I have not used this before, although I have used Comsol's general PDE module and other FEM libraries to solve the Cahn-Hilliard model. If we were to generalize to the Cahn-Hilliard/Navier Stokes model (at least, a specific version of it), it would not be very simple to impose general boundary conditions as you can with an immersed boundary method (continuous or discontinuous). There is a free energy that is minimized to formulate conservation of mass and momentum, the terms associated with the interface are derived indirectly, so you would need to find a free energy expression that is both physically relevant and results in the desired fluid/solid boundary conditions. With a numerical method for imposing boundary conditions at a fluid/solid interface, you can simply impose them directly without having to model the solid phase. Does that make sense?
