### 2022-03-09 uw-comphys/opencmp#6 @bonh
```
pip show configparser
Name: configparser
Version: 5.2.0
Summary: Updated configparser from Python 3.8 for Python 2.6+.
Home-page: https://github.com/jaraco/configparser/
Author: Łukasz Langa
Author-email: lukasz@langa.pl
License: UNKNOWN
Location: /usr/local/lib/python3.8/dist-packages
Requires:
Required-by:
```
