### 2022-03-09 uw-comphys/opencmp#6 @bonh
Just tried to open the results from `python3 -m opencmp config_IC` with `paraview 5.7.0` and it worked perfectly.
