### 2022-03-09 uw-comphys/opencmp#6 @nasserma
> Running the example works now. Here is the version info:
>
> ```
> pip show configparser
> Name: configparser
> Version: 5.2.0
> Summary: Updated configparser from Python 3.8 for Python 2.6+.
> Home-page: https://github.com/jaraco/configparser/
> Author: Łukasz Langa
> Author-email: lukasz@langa.pl
> License: UNKNOWN
> Location: /usr/local/lib/python3.8/dist-packages
> Requires:
> Required-by:
> ```

This seems to be a bug in the configparser module, using version 5.2.0 is treating the string "cyl" as a list for some reason, version 5.1.0 and previous versions do not do this (as expected). I will look into whether or not something changed in v5.2.0 that we need to account for, but it will take me a bit of time.

I think the same is the case with the Paraview issue, the .pvd file we create seems to have issues with v5.9 and higher.
