### 2022-03-09 uw-comphys/opencmp#6 @bonh
I am done here and fine with closing the issue.
