### 2022-03-15 #3742 @lucydot
Excellent @bonh thank-you very much for reviewing for JOSS.

@WilkAndy please could you give an update on your review?
