### 2022-03-15 #3742 @WilkAndy
Suggestion (not mandatory for publication): improve the lighting in your paraview figures - they're super dark at the moment!
