### 2022-03-30 #3742 @lucydot
@nasserma it looks like there may be an issue with an inline reference. The error message reads `attempt to get length of a nil value (local 'author')`.

Looking at the bib file, there are "author" fields missing in the "misc" entries, but I don't think these should be causing the problem. Perhaps you need double curly brackets around "Message Passing Interface Forum" as this is not in the standard author format .
