### 2022-03-30 #3742 @nasserma
@WilkAndy @lucydot I have improved the figures for tutorials 8 and 10 in the revised manuscript (above), the only issue left are that the figures for tutorial 9 need to be side-by-side. There seem to be a few ways to do this using Markdown, if there is a preferred one please let me know.
