### 2022-04-04 #3742 @lucydot
Hi @nasserma -  I can see the double brackets around "Message Passing Interface Forum" seemed to fix the problem  - great.

There is not a preferred method for aligning images in markdown that I am aware of. I think the best approach is to implement, then test that it compiles into pdf as expected here in the thread, using the editorialbot as you have been doing. It may be easier to have a single image file with two sub-images rather than aligning in markdown. The images do look rather large at the moment, I suggest that they can all be scaled down for easier readability of the paper as a whole.
