### 2022-04-13 #3742 @lucydot
@nasserma the latest version of the pdf has two of the images split over two pages. I'm not aware of a programmatic way to stop breaking images over two pages (I didn't know that markdown would allow this in fact) - perhaps you could try moving the position of the images within the main text so that this doesn't happen?
