### 2022-04-13 #3742 @lucydot
@nasserma a few points that remain from my previous comments:
- 'Nguyen et al. (2018)' should read '(Nguyen et al., 2018)' for consistency
- OpenFOAM and SU2 are linked to directly rather than referenced: can you confirm that references for each package are not available? If references are available I suggest you provide this in the text as done for the other packages.
- the arrow formatting has not worked: 'A&rarrB' is displayed.

Please check after making changes to the markdown that they appear as expected when viewing the compiled pdf. Thanks - Lucy.
