### 2022-04-19 #3742 @nasserma
@lucydot Please see the latest version which has the following changes,

- Two sets of images were merged so that they will not be split across pages (tutorials 8 and 10 images).
- Updated citation of '(Nguyen et al., 2018)', I was trying to cite to papers in a single citation.
- OpenFOAM and SU2 are cited per instructions from their websites
- arrow formatting corrected

Hopefully this resolves all of the issues!
