### 2022-04-21 #3742 @lucydot
@nasserma I'm just checking with other editorial team members about citatiion formatting when there is not a date associated with a publication (currently displaying as "n.d").
