### 2022-04-21 #3742 @lucydot
@nasserma
- the formatting for the MPI Forum author field is not correct - in your bib file the title should be in double curly brackets: "{{Message Parsing Interface Forum}}"
- there seems to be an inconsistency with the "misc" entries in the bib file. A `year` field has been supplied for ngsolve, but not for the other websites. I suggest removing the `year` field for ngsolve (which I think is a little vague in this context - year accessed or published?) and adding to all "misc" fields a `urldate` field in the following format `urldate={2014-04-01}`.
