### 2022-04-21 #3742 @nasserma
Hi @lucydot , I made the changes specified, but the adding the urldate field to the bib entries does not change the presence of "N.D." in the resulting citations. Also, I have populated that field with the date that the page was updated (extracted from the HTML metadata), but in the article proof it indicates the field should reflect the date accessed. I am pushing an update to the OpenFOAM bib entry now, since that was the only one that was not dynamically created.
