### 2022-04-25 #3742 @lucydot
Ah @nasserma thanks for trying the `urldate` field. Yes, the `urldate` is the date the page was accessed - if you could update accordingly (to the date accessed) I think we are then ready to recommend to editor in chiefs for acceptance. We'll leave the `n.d.` as it is - I can see there are other JOSS papers that have been published with this notation instead of a date.
