### 2022-05-05 #3742 @editorialbot
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
