### 2022-05-06 #3742 @nasserma
> Congratulations @nasserma on your article's publication in JOSS!
>
> Many thanks to @bonh and @WilkAndy for reviewing this, and @lucydot for editing.

Thanks to all!
