### 2022-09-21 #3742 @tarleb
Hi all! We are working towards a more unified archive of all JOSS papers. In this process, I noted that the Markdown sources for this paper contain markup to include an image that does not exist, namely `images/tutorial_8_geom.png`, and is hence not included in the PDF. Judging from the surrounding text, the image may have been combined with another graphic, so this may not be a problem. Otherwise we may want to re-publish this paper with the correct image. @nasserma, could you take a look?
