### 2022-09-25 #3742 @nasserma
Hi @tarleb , the situation is as you expected and line 106 of the current version of the Markdown source for the paper should be removed. I will push a revision of the file to the OpenCMP repository, but since this does not result in a change in the compiled PDF does anything else need to be done?
