### 2022-09-26 #3742 @tarleb
Thanks for the feedback. I'm generating JATS from a copy of the paper sources, so I can fix locally. With the PDF looking as intended, there is nothing else we'd need to do here. The rest is on my side. :+1:
